% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{alluvialClonotypes}
\alias{alluvialClonotypes}
\title{Exploring interaction of clonotypes by seurat dynamics}
\usage{
alluvialClonotypes(
  seurat,
  call = c("gene", "nt", "aa", "gene+nt"),
  compare = c("cluster", ...),
  facet = NULL
)
}
\arguments{
\item{seurat}{The seurat object to attach}

\item{call}{How to call the clonotype - based on genes, CDR3 nt, or CDR3 aa sequence or the combination of genes and nucleotide sequence}

\item{compare}{The column label to visualize}

\item{facet}{The column label to seperate}
}
\description{
Exploring interaction of clonotypes by seurat dynamics
}
