% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{alluvialGraph}
\alias{alluvialGraph}
\title{Exploring interaction of clonotypes by seurat dynamics}
\usage{
alluvialGraph(
  seurat,
  call = c("gene", "nt", "aa", "gene+nt"),
  compare = c("cluster", ...),
  facet = NULL
)
}
\arguments{
\item{seurat}{is the seurat object to attach}

\item{call}{is the how to call the clonotype - based on genes, CDR3 nt, or CDR3 aa sequence}

\item{compare}{is the column label to visualize}

\item{facet}{is the column label to seperate}
}
\description{
Exploring interaction of clonotypes by seurat dynamics
}
