% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\name{calIndex}
\alias{calIndex}
\alias{Startrac.calIndex}
\alias{calIndex,Startrac-method}
\title{Calculate cluster level indices}
\usage{
Startrac.calIndex(object, cores, n.perm, normEntropy)

\S4method{calIndex}{Startrac}(object, cores = NULL, n.perm = NULL, normEntropy = FALSE)
}
\arguments{
\item{object}{A Startrac object}

\item{cores}{number of core to be used. Passed to doParallel registerDoParallel. default: NULL.}

\item{n.perm}{integer number of permutation will be performed. If NULL, no permutation. (default: NULL)}

\item{normEntropy}{logical; whether normalize migration and transition index. default: FALSE.}
}
\value{
an object of class \code{Startrac}
}
\description{
Calculate cluster level indices
}
\keyword{internal}
