% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{changeNames}
\alias{changeNames}
\title{Change names of the cells in the Seurat object in order to use combineSeurat()}
\usage{
changeNames(seurat, seuratID = NULL, newID = NULL)
}
\arguments{
\item{seurat}{The Seurat object}

\item{seuratID}{The strings/prefixes of the current cell names to replace, i.e. c("PY_Tumor", "PX_Peripheral", "PZ_Tumor")}

\item{newID}{The strings to replace the Seurat cell names, i.e., c("PY_T", "PX_P", "PZ_T")}
}
\description{
Change the prefix of the seurat object in order to use combineSeurat().
}
