% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkContig.R
\name{checkContig}
\alias{checkContig}
\title{Assess the number of NAs in clonotypes}
\usage{
checkContig(
  df,
  chain = "TRA",
  examine = "gene",
  group.by = NULL,
  split.by = NULL,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{examine}{indicate either "chain" for sequence-based NA or 
"gene" for NA percentages across the VDJC genes.}

\item{group.by}{The column header used for grouping.}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{exportTable}{Returns the data frame used for forming the graph}

\item{palette}{Colors to use in visualization - input any hcl.pals()}
}
\value{
ggplot of the percentage of NA values
}
\description{
This function quantifies the percent of NA values in either the 
sequence information of the VDJC gene usage. If a data frame 
output for the raw numbers and percentage of NAs are preferred, 
set exportTable = TRUE.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
checkContig(combined, chain = "TRA", examine = "chain", group.by = "sample")

}
