% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{clonalOverlay}
\alias{clonalOverlay}
\title{Visualize distribution of clonal frequency overlaid on dimensional reduction plots}
\usage{
clonalOverlay(
  sc,
  reduction = NULL,
  freq.cutpoint = 30,
  bins = 25,
  facet = NULL
)
}
\arguments{
\item{sc}{The seurat or SCE object to visualize after combineExpression().}

\item{reduction}{The dimensional reduction to visualize}

\item{freq.cutpoint}{The overlay cutpoint to include, this corresponds to the 
Frequency variable in the single-cell objecter}

\item{bins}{The number of contours to the overlay}

\item{facet}{meta data variable to facet the comparison}
}
\value{
ggplot object
}
\description{
This function allows the user to visualize the clonal expansion by overlaying the 
cells with specific clonal frequency onto the dimensional reduction plots in Seurat.
Credit to the idea goes to Dr. Carmonia and his work with
[ProjectTIL](https://github.com/carmonalab/ProjecTILs).
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
sce <- suppressMessages(Seurat::UpdateSeuratObject(screp_example))
sce <- Seurat::as.SingleCellExperiment(sce)

#Using combineExpresion()
sce <- combineExpression(combined, sce)

#Using clonalOverlay()
clonalOverlay(sce, reduction = "umap", freq.cutpoint = 2, bins = 25, facet = "Patient") 

}
\author{
Francesco Mazziotta, Nick Borcherding
}
