% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{clonesizeDistribution}
\alias{clonesizeDistribution}
\title{Hierarchical clustering based on JS distance}
\usage{
clonesizeDistribution(
  df,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  method = c("complete", "ward.D", "ward.D2", "single", "average", "mcquitty",
    "median", "centroid"),
  exportTable = F
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{cloneCall}{How to call the clonotype - CDR3 gene, CDR3 nt or CDR3 aa, or CDR3+nucleotide}

\item{method}{The clustering paramater for the dendrogram}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\description{
Allows for the hierarchical clustering based on Jensen-Shannon distance using the discrete gamma-GPD spliced threshold model in the powerTCR R package. If you are planning on using this function for your analysis, read and cite PMID: 30485278.
}
