% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalBias.R
\name{clonotypeBias}
\alias{clonotypeBias}
\title{Examine clonotype bias}
\usage{
clonotypeBias(
  df,
  cloneCall = "strict",
  split.by = NULL,
  group.by = NULL,
  n.boots = 20,
  min.expand = 10,
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), 
or combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{split.by}{The column header used for calculating the baseline frequencies.
For example, "Type" for tumor vs peripheral blood comparison}

\item{group.by}{The column header used for comparisons of bias.}

\item{n.boots}{number of bootstraps to downsample}

\item{min.expand}{clonotype frequency cut off for the purpose of comparison}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\value{
Returns ggplot of the clonotype bias
}
\description{
Clonotype bias method was developed and outlined from a single-cell 
\href{https://pubmed.ncbi.nlm.nih.gov/35829695/}{manuscript} 
characterizing CD4 responses to acute and chronic infection. The metric seeks to 
quantify how individual clones are skewed towards a specific cellular 
compartment or cluster. A clonotype bias of 1 indicates that a clonotype 
is composed of cells from a single compartment or cluster, while a clonotype
bias of 0 matches the background subtype distribution.
}
\examples{
\dontrun{
Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
sce <- suppressMessages(Seurat::UpdateSeuratObject(screp_example))

#Using combineExpresion()
sce <- combineExpression(combined, sce)

#Using occupiedscRepertoire()
clonotypeBias(sce, cloneCall = "CTaa", split.by = "Patient", group.by = "seurat_clusters",
n.boots = 20, min.expand = 2)
}

}
