% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{combineContigs}
\alias{combineContigs}
\title{Combining the list of contigs}
\usage{
combineContigs(
  df,
  samples = NULL,
  ID = NULL,
  cells = c("T-AB", "T-GD", "B"),
  filterNA = F
)
}
\arguments{
\item{df}{List of filtered contig annotations from 10x Genomics}

\item{samples}{The labels of samples}

\item{ID}{The additional sample labeling option}

\item{cells}{The type of lymphocyte - T cell-AB, T cell-GD, or B cell}

\item{filterNA}{This will remove any chain without values}
}
\description{
Combining the list of contigs
}
