% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{combineSeurat}
\alias{combineSeurat}
\title{Adding clonotype information to a seurat object}
\usage{
combineSeurat(
  df,
  seurat,
  call = c("gene", "nt", "aa"),
  groupBy = c("none", "sample", "ID", ...),
  cloneTypes = c(Single = 1, Small = 5, Medium = 20, Large = 100, Hyperexpanded = 500)
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{seurat}{is the seurat object to attach}

\item{call}{is the how to call the clonotype - based on genes, CDR3 nt, or CDR3 aa sequence}

\item{groupBy}{is the column label in which clonotype frequency will be calculated}

\item{cloneType}{are the bins for the grouping based on frequency}
}
\description{
Adding clonotype information to a seurat object
}
