% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineTRUST4.R
\name{combineTRUST4}
\alias{combineTRUST4}
\title{Combining the list of T or B cell Receptors from TRUST4 pipeline}
\usage{
combineTRUST4(
  df,
  samples = NULL,
  ID = NULL,
  cells = c("T-AB", "T-GD", "B"),
  removeNA = FALSE,
  removeMulti = FALSE
)
}
\arguments{
\item{df}{List of Contig outputs from TRUST4}

\item{samples}{The labels of samples.}

\item{ID}{The additional sample labeling option.}

\item{cells}{The type of cell - T cell-AB or T cell-GD, or B cell}

\item{removeNA}{This will remove any chain without values.}

\item{removeMulti}{This will remove barcodes with greater than 2 chains.}
}
\value{
List of clonotypes for individual cell barcodes
}
\description{
This function consolidates a list of TCR/BCR sequencing results to the level of 
the individual cell barcodes using the same approach as 
 \code{\link{combineTCR}} and \code{\link{combineBCR}}. 
Using the samples and ID parameters, the function will add the strings 
as prefixes to prevent issues with repeated 
barcodes. The resulting new barcodes will need to match the seurat or SCE 
object in order to use, \code{\link{combineExpression}}. Several 
levels of filtering exist - removeMulti are parameters that 
control how  the function  deals with barcodes with multiple chains 
recovered. Please [read more](https://github.com/liulab-dfci/TRUST4) 
and cite the TRUST4 pipeline if using this function.
}
\details{

}
