% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{compareClonotypes}
\alias{compareClonotypes}
\title{Demonstrate the difference in clonal proportion between multiple clonotypes}
\usage{
compareClonotypes(
  df,
  call = c("gene", "nt", "aa", "gene+nt"),
  samples = NULL,
  clonotypes = NULL,
  numbers = NULL,
  graph = c("alluvial", "area")
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{call}{is the how to call the clonotype - CDR3 nt or CDR3 aa sequence.}

\item{clonotypes}{the specific sequences of interest}

\item{numbers}{the top n clonotype sequences}

\item{graph}{either "alluvial" or "area}
}
\description{
Demonstrate the difference in clonal proportion between multiple clonotypes
}
