% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportClones.R
\name{._bind_contig_list}
\alias{._bind_contig_list}
\title{Bind a List of Contig Data Frames and Add Grouping Variable}
\usage{
._bind_contig_list(contig_list)
}
\description{
Bind a List of Contig Data Frames and Add Grouping Variable
}
\keyword{internal}
