% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportClones.R
\name{._split_and_pad}
\alias{._split_and_pad}
\title{Split String and Pad to a Fixed-Width Matrix}
\usage{
._split_and_pad(x, split, n_cols)
}
\description{
Split String and Pad to a Fixed-Width Matrix
}
\keyword{internal}
