% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{expression2List}
\alias{expression2List}
\title{Allows users to take the meta data in seurat/SCE and place it into a list that will work with all the functions}
\usage{
expression2List(sc, group)
}
\arguments{
\item{sc}{object after combineExpression().}

\item{group}{The column header to group the new list by}
}
\description{
Allows users to perform more funadmental measures of clonotype analysis using the meta data from the seurat or SCE
object. For Seurat objects the active identity is automatically added as "cluster". Reamining grouping parameters
for SCE or Seurat objects must appear in the meta data.
}
