% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietVDJgenes.R
\name{findVariableNonVdjFeatures}
\alias{findVariableNonVdjFeatures}
\alias{findVariableNonVdjFeatures.Seurat}
\title{Find Variable Non-VDJ Features in a Seurat Object}
\usage{
findVariableNonVdjFeatures(sc, ...)

\method{findVariableNonVdjFeatures}{Seurat}(
  sc,
  assay = NULL,
  layer = NULL,
  finder = Seurat::FindVariableFeatures,
  filterVal = 0,
  ...
)
}
\arguments{
\item{sc}{A Seurat object.}

\item{...}{Additional arguments passed to the \code{finder} function.}

\item{assay}{The assay to use. Only supports assays that has RNA feature
names and the behaviour is otherwise undefined.}

\item{layer}{Character. The data layer to use, or NULL to use default data.}

\item{finder}{A function used to find variable features. Defaults to
\code{\link[Seurat:FindVariableFeatures]{Seurat::FindVariableFeatures()}}.}

\item{filterVal}{The value to set the filtered genes to. Defaults to 0.}
}
\value{
A modified Seurat object with updated variable features.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function identifies variable features from a specified assay of a Seurat
object, excluding features associated with V(D)J genes. It is useful for
preprocessing RNA assays in single-cell datasets where TCR or BCR genes
should not be considered like clustering so that the clustering is not
biased by phenotype.

This is an improvement to first finding variable features then using
\code{\link[=quietVDJgenes]{quietVDJgenes()}} to filter out variable TCR and BCR genes. This instead
does the filtering ahead of time and then finds variable features after,
resulting potentially in a more accurate selection. Note that filtering
in this context is simply setting all the V(D)J genes to a constant value (0)
so they are not considered during variable feature selection.
}
\details{
Currently, this function may be somewhat performance intensive as it
creates a temporary copy of the data in the filtering step. Note that
it also requires the Seurat package to be loaded at least via namespace.
}
\examples{
findVariableNonVdjFeatures(scRep_example, nfeatures = 2000)

}
