% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quietVDJgenes.R
\name{getHumanIgPseudoGenes}
\alias{getHumanIgPseudoGenes}
\title{Get Human Immunoglobulin pseudogenes}
\usage{
getHumanIgPseudoGenes()
}
\value{
Character vector of human immunoglobulin pseudogenes.
}
\description{
This function returns a character vector of human immunoglobulin
pseudogenes. These are also the genes that are removed from the
variable gene list in \code{\link[=quietVDJgenes]{quietVDJgenes()}}.
}
\keyword{internal}
