% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\name{getSig}
\alias{getSig}
\alias{Startrac.getSig}
\alias{getSig,Startrac-method}
\title{Get the p value given one Startrac object and a list of Startrac objects from permutation data}
\usage{
Startrac.getSig(obj, obj.perm)

\S4method{getSig}{Startrac}(obj, obj.perm = NULL)
}
\arguments{
\item{obj}{A Startrac object}

\item{obj.perm}{A list of Startrac objects from permutation data}
}
\value{
an object of class \code{Startrac}
}
\description{
Get the p value given one Startrac object and a list of Startrac objects from permutation data
}
\keyword{internal}
