% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{hammingCompare}
\alias{hammingCompare}
\title{Calculates the normalized Hamming Distance between the nucleotide sequence of the heavy or light chain. Relies on the length of the nucleotide sequence for comparison.}
\usage{
hammingCompare(gene, chain, length)
}
\arguments{
\item{gene}{The IGH or IG light chains (IGLC)}

\item{chain}{The column header with the nucletoide sequence}

\item{length}{The column header with the specific length}
}
\description{
Calculates the normalized Hamming Distance between the nucleotide sequence of the heavy or light chain. Relies on the length of the nucleotide sequence for comparison.
}
