% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{initialize}
\alias{initialize}
\alias{initialize,Startrac-method}
\title{initialize method for Startrac}
\usage{
\S4method{initialize}{Startrac}(.Object, cell.data, aid = "AID", n.perm = NULL, cores = NULL)
}
\arguments{
\item{.Object}{A Startrac object}

\item{cell.data}{data.frame contains the input data}

\item{aid}{character analysis id}

\item{n.perm}{integer number of permutation will be performed. If NULL, no permutation. (default: NULL)}

\item{cores}{number of core to be used. Passed to doParallel registerDoParallel. default: NULL.}
}
\value{
an object of class \code{Startrac}
}
\description{
initialize method for Startrac
}
\keyword{internal}
