% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{initialize,StartracOut-method}
\alias{initialize,StartracOut-method}
\title{initialize method for StartracOut}
\usage{
\S4method{initialize}{StartracOut}(.Object, proj = "AID")
}
\arguments{
\item{.Object}{A StartracOut object}

\item{proj}{character analysis id}
}
\value{
an object of class \code{StartracOut}
}
\description{
initialize method for StartracOut
}
\keyword{internal}
