% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{lengthContig}
\alias{lengthContig}
\title{Demonstrate the distribution of lengths filtered contigs output from 10x Genomics}
\usage{
lengthContig(
  df,
  cloneCall = c("nt", "aa"),
  column = NULL,
  scale = F,
  chains = c("combined", "single"),
  exportTable = F
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{cloneCall}{How to call the clonotype - CDR3 nt or CDR3 aa sequence.}

\item{column}{The column header for which you would like to analyze the data}

\item{scale}{Converts the graphs into denisty plots in order to show relative distributions.}

\item{chains}{Whether to keep clonotypes "combined" or visualize by chain}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\description{
Examine eithe the nucleotide (nt) or amino acid (aa) sequence length across samples or by vairable (column). Sequences can be visualized as combined values (both chains), or as single chains. If more than 2 chains of the same locus are assigned to the individual barcode, this function will take the first chain.
}
