% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadContigs.R
\name{loadContigs}
\alias{loadContigs}
\title{Loading the contigs derived from single-cell sequencing}
\usage{
loadContigs(dir, format = "10X")
}
\arguments{
\item{dir}{The directory or directories with single-cell contig data}

\item{format}{The format of the single-cell contig, currently supporting: "10X", "AIRR", "TRUST4", and "WAT3R"}
}
\value{
List of contigs for further processing in scRepertoire
}
\description{
This function generates a contig list and formats the data to allow for 
function with  \code{\link{combineTCR}} or \code{\link{combineTCR}}. If 
using data derived from filtered outputs of 10X Genomics, there is no 
need to use this function as the data is already compatible. The function
assumes if listing multiple directories, there are distinct outputs with
unmodified file names in them.
}
\examples{
\dontrun{
dir <- c("./Sample1/outs/", "./Sample2/outs/", "./Sample3/outs/")
contig.list <- loadContigs(dir, format = "10X")
}

}
