% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\name{loginfo}
\alias{loginfo}
\title{dispaly message with time stamp}
\usage{
loginfo(msg)
}
\arguments{
\item{msg}{characters; message to display}
}
\value{
Estimate of sys.time
}
\description{
dispaly message with time stamp
}
\keyword{internal}
