% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\name{mcol.entropy}
\alias{mcol.entropy}
\title{entropy of each column of the input matrix}
\usage{
mcol.entropy(x)
}
\arguments{
\item{x}{matrix;}
}
\value{
column entropy
}
\description{
entropy of each column of the input matrix
}
\keyword{internal}
