% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_contig_list}
\alias{mini_contig_list}
\title{Processed subset of \code{contig_list}}
\format{
An R \code{list} of \code{data.frame} objects
}
\usage{
data("mini_contig_list")
}
\description{
A list of 8 data frames of T cell contigs outputted from the
\code{filtered_contig_annotation} files, but subsetted to  365 valid T cells
which correspond to the same barcodes found in \code{scRep_example}. The
data is originally derived from the following
\href{https://pubmed.ncbi.nlm.nih.gov/33622974/}{manuscript}.
}
\seealso{
\code{\link[=contig_list]{contig_list()}}
}
\concept{Data}
\keyword{datasets}
