% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\name{mrow.entropy}
\alias{mrow.entropy}
\title{entropy of each row of the input matrix}
\usage{
mrow.entropy(x)
}
\arguments{
\item{x}{matrix;}
}
\value{
row entropy
}
\description{
entropy of each row of the input matrix
}
\keyword{internal}
