% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{occupiedscRepertoire}
\alias{occupiedscRepertoire}
\title{Visualize the number of single cells with clonotype frequencies by cluster}
\usage{
occupiedscRepertoire(sc, x.axis = "cluster", exportTable = FALSE)
}
\arguments{
\item{sc}{The seurat or SCE object to visualize after combineExpression(). 
For SCE objects, the cluster variable must be in the meta data under 
"cluster".}

\item{x.axis}{The variable in the meta data to graph along the x.axis}

\item{exportTable}{Exports a table of the data into the global 
environment in addition to the visualization}
}
\value{
Stacked bar plot of counts of cells by clonotype frequency group
}
\description{
View the count of clonotypes frequency group in seurat or SCE object 
meta data after combineExpression(). The visualization will take the 
new meta data variable "cloneType" and plot the number of cells with
each designation using a secondary variable, like cluster. Credit to 
the idea goes to Dr. Carmonia and his work with
[ProjectTIL](https://github.com/carmonalab/ProjecTILs).
}
