% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentGenes.R
\name{percentGenes}
\alias{percentGenes}
\title{Examining the VDJ gene usage across clones}
\usage{
percentGenes(
  input.data,
  chain = "TRB",
  gene = "Vgene",
  group.by = NULL,
  order.by = NULL,
  exportTable = FALSE,
  summary.fun = c("percent", "proportion", "count"),
  palette = "inferno"
)
}
\arguments{
\item{input.data}{The product of \code{\link[=combineTCR]{combineTCR()}},
\code{\link[=combineBCR]{combineBCR()}}, or \code{\link[=combineExpression]{combineExpression()}}.}

\item{chain}{indicate a specific chain should be used -
e.g. "TRA", "TRG", "IGH", "IGL", etc}

\item{gene}{"V", "D" or "J"}

\item{group.by}{The variable to use for grouping}

\item{order.by}{A vector of specific plotting order or "alphanumeric"
to plot groups in order}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{summary.fun}{Character string choosing the summary statistic -
\code{"percent"} (default), \code{"proportion"}, or \code{"count"}}

\item{palette}{Colors to use in visualization - input any
\link[grDevices:palettes]{hcl.pals}.}
}
\value{
A ggplot object displaying a heatmap of gene percentages.
If \code{exportTable = TRUE}, a matrix of the raw data is returned.
}
\description{
This function the proportion, percent, or counts of V, D, or J genes used by
grouping variables. This function only quantifies single gene loci for
indicated \strong{chain}. For examining VJ or other gene parings, please
see \code{\link[=percentVJ]{percentVJ()}} or \code{\link[=vizGenes]{vizGenes()}}.
}
\examples{
# Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
percentGenes(combined, 
             chain = "TRB", 
             gene = "Vgene")

}
\concept{Summarize_Repertoire}
