% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{quantContig}
\alias{quantContig}
\title{Quantify the unique clonotypes in the filtered contigs output from 10x Genomics}
\usage{
quantContig(
  df,
  call = c("gene", "nt", "aa", "gene+nt"),
  scale = F,
  column = NULL
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{call}{is the how to call the clonotype - CDR3 gene, CDR3 nt or CDR3 aa, or CDR3+nucleotide}

\item{scale}{converts the graphs into percentage of unique clonotypes}

\item{column}{is the column header for which you would like to analyze the data}
}
\description{
Quantify the unique clonotypes in the filtered contigs output from 10x Genomics
}
