% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{scatterClonotype}
\alias{scatterClonotype}
\title{Scatter plot comparing the expansion of two samples
This function produces a scatter plot directly comparing the specific clonotypes
between two samples. The clonotypes will be categorized by counts into singlets or multiplets, 
either exclusive or shared between the selected samples. Visualization inspired 
by the work of \href{https://pubmed.ncbi.nlm.nih.gov/32103181/}{Wu, T, et al 2020}.}
\usage{
scatterClonotype(
  df,
  cloneCall = "gene+nt",
  x.axis = NULL,
  y.axis = NULL,
  graph = "proportion",
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), or expression2List()}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
CDR3 gene+nucleotide (gene+nt).}

\item{x.axis}{name of the list element to appear on the x.axis}

\item{y.axis}{name of the list element to appear on the y.axis}

\item{graph}{graph either proportion or raw clonotype count}

\item{exportTable}{Returns the data frame used for forming the graph.}
}
\value{
ggplot of the relative clonotype numbers
}
\description{
Scatter plot comparing the expansion of two samples
This function produces a scatter plot directly comparing the specific clonotypes
between two samples. The clonotypes will be categorized by counts into singlets or multiplets, 
either exclusive or shared between the selected samples. Visualization inspired 
by the work of \href{https://pubmed.ncbi.nlm.nih.gov/32103181/}{Wu, T, et al 2020}.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")
compareClonotypes(combined, numbers = 10, 
samples = c("PX_P", "PX_T"), cloneCall="aa")

}
