% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{screp_example}
\alias{screp_example}
\title{A seurat object of 1000 single T cells derived
from 3 clear cell renal carcinoma patients.}
\description{
A seurat object of 1000 single T cells derived
from 3 clear cell renal carcinoma patients.
}
