% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRepertoire-package.R
\docType{package}
\name{scRepertoire-package}
\alias{scRepertoire}
\alias{scRepertoire-package}
\title{scRepertoire: A toolkit for single-cell immune receptor profiling}
\description{
scRepertoire was built to process data derived from the 10x Genomics Chromium Immune Profiling for both T-cell receptor (TCR) and immunoglobulin (Ig) enrichment workflows and subsequently interacts with the popular Seurat and SingleCellExperiment R packages.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ncborcherding.github.io/scRepertoire/}
}

}
\author{
\strong{Maintainer}: Nick Borcherding \email{ncborch@gmail.com}

Authors:
\itemize{
  \item Qile Yang \email{qile0317@gmail.com}
}

}
\keyword{internal}
