% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{seurat2List}
\alias{seurat2List}
\title{Allows users to take the meta data in seurat and place it into a list that will work with all the functions}
\usage{
seurat2List(df)
}
\arguments{
\item{df}{seurat object after combineSeurat()}
}
\description{
Allows users to take the meta data in seurat and place it into a list that will work with all the functions
}
