% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{show,StartracOut-method}
\alias{show,StartracOut-method}
\title{show method for StartracOut}
\usage{
\S4method{show}{StartracOut}(object)
}
\arguments{
\item{object}{A StartracOut object}
}
\description{
show method for StartracOut
}
\keyword{internal}
