% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,Startrac-method}
\title{show method for Startrac}
\usage{
\S4method{show}{Startrac}(object)
}
\arguments{
\item{object}{A Startrac object}
}
\value{
method for show
}
\description{
show method for Startrac
}
\keyword{internal}
