% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{stripBarcode}
\alias{stripBarcode}
\title{Removing any additional prefixes to the barcodes of the filtered contigs.}
\usage{
stripBarcode(contigs, column = 1, connector = "_", num_connects = 3)
}
\arguments{
\item{contigs}{raw loaded filtered_contig_annotation.csv}

\item{column}{is column in which the barcodes are listed}

\item{connector}{is the type of character in which is attaching the defualt barcode with any other characters}

\item{num_of_connects}{is the number of strings combined with the connectors}
}
\description{
Removing any additional prefixes to the barcodes of the filtered contigs.
}
