% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{vizGenes}
\alias{vizGenes}
\title{Visualizing the distribution of gene usage}
\usage{
vizGenes(
  df,
  gene = "V",
  chain = "TRA",
  plot = "heatmap",
  separate = "sample",
  order = "gene",
  scale = TRUE,
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{gene}{Which part of the immune receptor to visualize - V, D, J, C}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "TRA", "TRG", "IGH", "IGL" (no both option here)}

\item{plot}{The type of plot to return - heatmap or bar}

\item{separate}{Categorical variable which to separate by along y-axis}

\item{order}{Categorixal variable to organize the x-axis, either "gene" or "variance"}

\item{scale}{Converts the proportion of total genes}

\item{exportTable}{Returns the data frame used for forming the graph.}
}
\value{
ggplot bar diagram of vgene counts
}
\description{
This function will allow for the visualizing the distribution 
of the any VDJ and C gene of the TCR or BCR using heatmap or
bar chart. This function requires assumes two chains were used in 
defining clonotype, if not, it will default to the only chain 
present regardless of the chain parameter.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

vizGenes(combined, gene = "V", chain = "TRB", plot = "bar", scale = TRUE)

}
