% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{vizVgenes}
\alias{vizVgenes}
\title{Visualizing the distribution of TCR V gene usage}
\usage{
vizVgenes(
  df,
  TCR = "TCR1",
  facet.x = "sample",
  facet.y = NULL,
  fill = NULL,
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), or expression2List().}

\item{TCR}{Which TCR chain to use, TCR1 = TCRA or TCR2 = TCRB}

\item{facet.x}{Categorical variable which to seperate by along x-axis}

\item{facet.y}{Categorical variable which to seperate by along y-axis}

\item{fill}{Categorical variable which to add color to bar chart}

\item{exportTable}{Returns the data frame used for forming the graph.}
}
\value{
ggplot bar diagram of vgene counts
}
\description{
This function will allow for the visualizing the distribution 
ofthe V-genes of the TCR by categroical variables.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

vizVgenes(combined, TCR = "TCR1", facet.x = "sample")

}
