% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{addVariable}
\alias{addVariable}
\title{Adding variables after the combination of contigs.}
\usage{
addVariable(df, name = NULL, variables = NULL)
}
\arguments{
\item{df}{The product of CombineContig().}

\item{name}{The column header to add.}

\item{variables}{The exact values to add to each element of the list.}
}
\value{
list of contigs with a new column (name).
}
\description{
This function adds variables to the product of combineContig() to be used in later visualizations. For each element
in the combinContig(), the function will add a column (labled by name) with the variable. The length of the
variable paramater needs to match the length of the combineContig() object.
}
\examples{
combined <- addVariable(combined, name = "batch", variables = c(1,1,2,3,3))

}
