% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addVariable.R
\name{addVariable}
\alias{addVariable}
\title{Adding variables after the combination of clonotypes.}
\usage{
addVariable(df, name = NULL, variables = NULL)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}} or \code{\link{combineBCR}}.}

\item{name}{The column header to add.}

\item{variables}{The exact values to add to each element of the list.}
}
\value{
list of contigs with a new column (name).
}
\description{
This function adds variables to the product of \code{\link{combineTCR}}, 
or \code{\link{combineBCR}} to be used in later visualizations. 
For each element, the function will add a column (labeled by name) 
with the variable. The length of the variable paramater needs to match
the length of the combined object.
}
\examples{
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
combined <- addVariable(combined, 
                       name = "Type", 
                       variables = rep(c("B", "L"), 4))

}
