% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{addVariable}
\alias{addVariable}
\title{Adding variables after the combination of contigs.}
\usage{
addVariable(df, name = NULL, variables = NULL)
}
\arguments{
\item{df}{The product of combineTCR() combineBCR() or expression2List().}

\item{name}{The column header to add.}

\item{variables}{The exact values to add to each element of the list.}
}
\value{
list of contigs with a new column (name).
}
\description{
This function adds variables to the product of combineTCR() combineBCR() or 
expression2List() to be used in later visualizations. For each element, 
the function will add a column (labled by name) with the variable. 
The length of the variable paramater needs to match the length of 
the combined object.
}
\examples{
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))
combined <- addVariable(combined, name = "batch", variables = c(1,1,1,1,2,2))

}
