% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkContig.R
\name{checkContig}
\alias{checkContig}
\title{Assess the number of NAs in clonotypes}
\usage{
checkContig(
  df,
  chain = "TRA",
  examine = "gene",
  group.by = NULL,
  split.by = NULL,
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{examine}{indicate either "chain" for sequence-based NA or 
"gene" for NA percentages across the VDJC genes.}

\item{group.by}{The column header used for grouping.}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\value{
ggplot of the percentage of NA values
}
\description{
This function takes the output from combineTCR(), combineBCR(), 
combineExpression() or expression2List() and quantifies the 
percent of NA values in either the sequence information of 
the VDJC gene usage. If a data frame output for the raw numbers 
and percentage of NAs are preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))
checkContig(combined, chain = "TRA", examine = "chain", group.by = "sample")

}
