% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceContig.R
\name{clonalAbundance}
\alias{clonalAbundance}
\title{Demonstrate the relative abundance of clonotypes by group or sample.}
\usage{
clonalAbundance(
  df,
  cloneCall = "strict",
  chain = "both",
  scale = FALSE,
  group.by = NULL,
  order = TRUE,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{scale}{Converts the graphs into density plots in order to show 
relative distributions.}

\item{group.by}{The variable to use for grouping}

\item{order}{Maintain the order of the list when plotting}

\item{exportTable}{Returns the data frame used for forming the graph
to the visualization.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}
}
\value{
ggplot of the total or relative abundance of clonotypes 
across quanta
}
\description{
Displays the number of clonotypes at specific requencies by sample 
or group. Visualization can either be a line graph using
calculated numbers or if scale = TRUE, the output will be a 
density plot. Multiple sequencing runs can be group together 
using the group parameter. If a matrix output for the data is 
preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalAbundance(combined, cloneCall = "gene", scale = FALSE)

}
