% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalBias.R
\name{clonalBias}
\alias{clonalBias}
\title{Examine clonotype bias}
\usage{
clonalBias(
  df,
  cloneCall = "strict",
  split.by = NULL,
  group.by = NULL,
  n.boots = 20,
  min.expand = 10,
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{split.by}{The variable to use for calculating the baseline frequencies.
For example, "Type" for lung vs peripheral blood comparison}

\item{group.by}{The variable to use for calculating bias}

\item{n.boots}{number of bootstraps to downsample}

\item{min.expand}{clonotype frequency cut off for the purpose of comparison}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\value{
Returns ggplot of the clonotype bias
}
\description{
The metric seeks to 
quantify how individual clones are skewed towards a specific cellular 
compartment or cluster. A clonotype bias of *1* - indicates that a clonotype 
is composed of cells from a single compartment or cluster, while a clonotype
bias of *0* - matches the background subtype distribution. Please read and cite 
the following
\href{https://pubmed.ncbi.nlm.nih.gov/35829695/}{manuscript} 
if using \code{\link{clonalBias()}}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

#Getting a sample of a Seurat object
scRep_example <- get(data("scRep_example"))

#Using combineExpresion()
scRep_example <- combineExpression(combined, scRep_example)
scRep_example$Patient <- substring(scRep_example$orig.ident,1,3)

#Using clonalBias()
clonalBias(scRep_example, 
              cloneCall = "aa", 
              split.by = "Patient", 
              group.by = "seurat_clusters",
              n.boots = 20, 
              min.expand = 2)


}
