% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalDiversity.R
\name{clonalDiversity}
\alias{clonalDiversity}
\title{Calculate the clonal diversity for samples or groupings}
\usage{
clonalDiversity(
  input.data,
  cloneCall = "strict",
  chain = "both",
  group.by = NULL,
  order.by = NULL,
  x.axis = NULL,
  metrics = c("shannon", "inv.simpson", "norm.entropy", "gini.simpson", "chao1", "ACE"),
  exportTable = FALSE,
  palette = "inferno",
  n.boots = 100,
  return.boots = FALSE,
  skip.boots = FALSE
)
}
\arguments{
\item{input.data}{The product of \code{\link[=combineTCR]{combineTCR()}},
\code{\link[=combineBCR]{combineBCR()}}, or \code{\link[=combineExpression]{combineExpression()}}.}

\item{cloneCall}{How to call the clone - VDJC gene (\strong{gene}),
CDR3 nucleotide (\strong{nt}), CDR3 amino acid (\strong{aa}),
VDJC gene + CDR3 nucleotide (\strong{strict}) or a custom variable
in the data}

\item{chain}{indicate if both or a specific chain should be used -
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{Variable in which to combine for the diversity calculation}

\item{order.by}{A vector of specific plotting order or "alphanumeric"
to plot groups in order}

\item{x.axis}{Additional variable grouping that will space the sample along the x-axis}

\item{metrics}{The indices to use in diversity calculations -
"shannon", "inv.simpson", "norm.entropy", "gini.simpson", "chao1", "ACE"}

\item{exportTable}{Exports a table of the data into the global environment
in addition to the visualization}

\item{palette}{Colors to use in visualization - input any
\link[grDevices:palettes]{hcl.pals}}

\item{n.boots}{number of bootstraps to down sample in order to
get mean diversity}

\item{return.boots}{export boot strapped values calculated -
will automatically exportTable = TRUE.}

\item{skip.boots}{remove down sampling and boot strapping from the calculation.}
}
\value{
ggplot of the diversity of clones by group
}
\description{
This function calculates traditional measures of diversity - \strong{Shannon},
\strong{inverse Simpson}, \strong{normalized entropy}, \strong{Gini-Simpson}, \strong{Chao1 index}, and
\strong{abundance-based coverage estimators (ACE)} measure of species evenness by sample or group.
The function automatically down samples the diversity metrics using
100 boot straps (\strong{n.boots = 100}) and outputs the mean of the values.
The group parameter can be used to condense the individual
samples. If a matrix output for the data is preferred, set \strong{exportTable} = TRUE.
}
\details{
The formulas for the indices and estimators are as follows:

\strong{Shannon Index:}
\deqn{Index = - \sum p_i * \log(p_i)}

\strong{Inverse Simpson Index:}
\deqn{Index = \frac{1}{(\sum_{i=1}^{S} p_i^2)}}

\strong{Normalized Entropy:}
\deqn{Index = -\frac{\sum_{i=1}^{S} p_i \ln(p_i)}{\ln(S)}}

\strong{Gini-Simpson Index:}
\deqn{Index = 1 - \sum_{i=1}^{S} p_i^2}

\strong{Chao1 Index:}
\deqn{Index = S_{obs} + \frac{n_1(n_1-1)}{2*n_2+1}}

\strong{Abundance-based Coverage Estimator (ACE):}
\deqn{Index = S_{abund} + \frac{S_{rare}}{C_{ace}} + \frac{F_1}{C_{ace}}}

Where:
\itemize{
\item{\eqn{p_i}{p[i]} is the proportion of species \eqn{i}{i} in the dataset.}
\item{\eqn{S}{S} is the total number of species.}
\item{\eqn{n_1}{n[1]} and \eqn{n_2}{n[2]} are the number of singletons and doubletons, respectively.}
\item{\eqn{S_{abund}}{S[abund]}, \eqn{S_{rare}}{S[rare]}, \eqn{C_{ace}}{C[ace]}, and \eqn{F_1}{F[1]} are parameters derived from the data.}
}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalDiversity(combined, cloneCall = "gene")

}
\author{
Andrew Malone, Nick Borcherding
}
\concept{Visualizing_Clones}
