% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{clonalDiversity}
\alias{clonalDiversity}
\title{Examine the clonal diversity of samples}
\usage{
clonalDiversity(
  df,
  call = c("gene", "nt", "aa", "gene+nt"),
  colorBy = c("samples", ...)
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{call}{How to call the clonotype - CDR3 gene, CDR3 nt or CDR3 aa, or CDR3+nucleotide}

\item{colorBy}{The column header for which you would like to analyze the data}
}
\description{
Examine the clonal diversity of samples
}
