% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalDiversity.R
\name{clonalDiversity}
\alias{clonalDiversity}
\title{Examine the clonal diversity of samples}
\usage{
clonalDiversity(
  df,
  cloneCall = "strict",
  chain = "both",
  group.by = NULL,
  x.axis = NULL,
  metrics = c("shannon", "inv.simpson", "norm.entropy", "gini.simpson", "chao1", "ACE"),
  exportTable = FALSE,
  palette = "inferno",
  n.boots = 100,
  return.boots = FALSE,
  skip.boots = FALSE
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL".}

\item{group.by}{The variable to use for grouping.}

\item{x.axis}{Additional variable in which to split the x.axis.}

\item{metrics}{The indices to use in diversity calculations - "shannon", "inv.simpson", 
"norm.entropy", "gini.simpson", "chao1", "ACE".}

\item{exportTable}{Exports a table of the data into the global environment 
in addition to the visualization.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}

\item{n.boots}{number of bootstraps to downsample in order to get mean diversity}

\item{return.boots}{export boot strapped values calculated - 
will automatically exportTable = TRUE.}

\item{skip.boots}{remove downsampling and boot strapping from the calculation.}
}
\value{
ggplot of the diversity of clones by group
}
\description{
This function calculates traditional measures of diversity - **Shannon**, 
**inverse Simpson**, **normalized entropy**, **Gini-Simpson**, **Chao1 index**, and
**abundance-based coverage estimators (ACE)** measure of species evenness by sample or group. 
The function automatically down samples the diversity metrics using 
100 boot straps The group parameter can be used to condense the individual 
samples. If a matrix output for the data is preferred, set exportTable = TRUE.
}
\details{
The formulas for the indices and estimators are as follows:

Shannon Index:
\deqn{H = - \sum p_i \cdot \log(p_i)}

Inverse Simpson Index:
\deqn{ D^{-1} = 1 / \sum p_i^2}

Normalized Entropy:
\deqn{E^H = H / \log(S)}

Gini-Simpson Index:
\deqn{1 - D = 1 - \sum p_i^2}

Chao1 Index:
\deqn{\hat{S}_{Chao1} = S + \left( \frac{n_1(n_1 - 1)}{2(n_2 + 1)} \right)}

Abundance-based Coverage Estimator (ACE): 
\deqn{\hat{S}_{ACE} = S_{abundant} + \frac{S_{rare}}{C_{rare}} + \left( \frac{S_{rare} - 1}{C_{rare}} \right) \cdot F_1}

Where:
\itemize{
  \item{\(p_i\) is the proportion of species \(i\) in the dataset.}
  \item{\(S\) is the total number of species.}
  \item{\(n_1\) and \(n_2\) are the number of singletons and doubletons, respectively.}
  \item{\(S_{abundant}\), \(S_{rare}\), \(C_{rare}\), and \(F_1\) are parameters derived from the data.}
}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalDiversity(combined, cloneCall = "gene")

}
\author{
Andrew Malone, Nick Borcherding
}
