% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{clonalDiversity}
\alias{clonalDiversity}
\title{Examine the clonal diversity of samples}
\usage{
clonalDiversity(
  df,
  cloneCall = "gene+nt",
  chain = "both",
  group = "samples",
  exportTable = FALSE,
  n.boots = 100
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(),  or expression2List().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (gene+nt).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group}{The column header for which you would like to analyze the data.}

\item{exportTable}{Exports a table of the data into the global environment 
in addition to the visualization}

\item{n.boots}{number of bootstraps to downsample in order to get mean diversity}
}
\value{
ggplot of the diversity of clonotype sequences across list
}
\description{
This function calculates traditional measures of diversity - Shannon, 
inverse Simpson, Chao1 index, and abundance-based coverage estimators 
(ACE) by sample or group. The function automatically down samples the
diversity metrics using 100 boot straps The group paramter can be 
used to condense the individual samples. If a matrix output for 
the data is preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")
clonalDiversity(combined, cloneCall = "gene")

}
\author{
Andrew Malone, Nick Borcherding
}
