% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity.R
\name{clonalDiversity}
\alias{clonalDiversity}
\title{Examine the clonal diversity of samples}
\usage{
clonalDiversity(
  df,
  cloneCall = "strict",
  chain = "both",
  group.by = NULL,
  x.axis = NULL,
  split.by = NULL,
  exportTable = FALSE,
  n.boots = 100,
  return.boots = FALSE,
  skip.boots = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{Variable in which to group the diversity calculation}

\item{x.axis}{Additional variable in which to split the x.axis}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{exportTable}{Exports a table of the data into the global environment 
in addition to the visualization}

\item{n.boots}{number of bootstraps to downsample in order to get mean diversity}

\item{return.boots}{export boot strapped values calculated - 
will automatically exportTable = TRUE}

\item{skip.boots}{remove downsampling and boot strapping from the calculation}
}
\value{
ggplot of the diversity of clonotype sequences across list
}
\description{
This function calculates traditional measures of diversity - Shannon, 
inverse Simpson, Chao1 index, abundance-based coverage estimators 
(ACE), and 1-Pielou's measure of species evenness by sample or group. 
The function automatically down samples the
diversity metrics using 100 boot straps The group parameter can be 
used to condense the individual samples. If a matrix output for 
the data is preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))
clonalDiversity(combined, cloneCall = "gene")

}
\author{
Andrew Malone, Nick Borcherding
}
