% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalHomeostasis.R
\name{clonalHomeostasis}
\alias{clonalHomeostasis}
\title{Examining the clonal homeostasis of the repertoire}
\usage{
clonalHomeostasis(
  input.data,
  cloneSize = c(Rare = 1e-04, Small = 0.001, Medium = 0.01, Large = 0.1, Hyperexpanded =
    1),
  cloneCall = "strict",
  chain = "both",
  group.by = NULL,
  order.by = NULL,
  exportTable = FALSE,
  palette = "inferno",
  ...
)
}
\arguments{
\item{input.data}{The product of \code{\link[=combineTCR]{combineTCR()}},
\code{\link[=combineBCR]{combineBCR()}}, or \code{\link[=combineExpression]{combineExpression()}}.}

\item{cloneSize}{The cut points of the proportions.}

\item{cloneCall}{How to call the clone - VDJC gene (\strong{gene}),
CDR3 nucleotide (\strong{nt}), CDR3 amino acid (\strong{aa}),
VDJC gene + CDR3 nucleotide (\strong{strict}) or a custom variable
in the data.}

\item{chain}{indicate if both or a specific chain should be used -
e.g. "both", "TRA", "TRG", "IGH", "IGL".}

\item{group.by}{The variable to use for grouping}

\item{order.by}{A vector of specific plotting order or "alphanumeric"
to plot groups in order}

\item{exportTable}{Exports a table of the data into the global
environment in addition to the visualization.}

\item{palette}{Colors to use in visualization - input any
\link[grDevices:palettes]{hcl.pals}.}

\item{...}{Additional arguments passed to the ggplot theme}
}
\value{
ggplot of the space occupied by the specific proportion of clones
}
\description{
This function calculates the space occupied by clone proportions.
The grouping of these clones is based on the parameter \strong{cloneSize},
at default, \strong{cloneSize} will group the clones into bins of Rare = 0
to 0.0001, Small = 0.0001 to 0.001, etc. To adjust the proportions,
change the number or labeling of the cloneSize parameter. If a matrix
output for the data is preferred, set \strong{exportTable} = TRUE.
}
\examples{
# Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalHomeostasis(combined, cloneCall = "gene")

}
\concept{Visualizing_Clones}
