% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalLength.R
\name{clonalLength}
\alias{clonalLength}
\title{Demonstrate the distribution of clonal length}
\usage{
clonalLength(
  input.data,
  cloneCall = "aa",
  chain = "both",
  group.by = NULL,
  order.by = NULL,
  scale = FALSE,
  exportTable = FALSE,
  palette = "inferno",
  ...
)
}
\arguments{
\item{input.data}{The product of \code{\link[=combineTCR]{combineTCR()}},
\code{\link[=combineBCR]{combineBCR()}}, or \code{\link[=combineExpression]{combineExpression()}}}

\item{cloneCall}{How to call the clone - CDR3 nucleotide (\strong{nt})
or CDR3 amino acid (\strong{aa})}

\item{chain}{indicate if both or a specific chain should be used -
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The variable to use for grouping}

\item{order.by}{A vector of specific plotting order or "alphanumeric"
to plot groups in order description}

\item{scale}{Converts the graphs into density plots in order to show
relative distributions.}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any
\link[grDevices:palettes]{hcl.pals}}

\item{...}{Additional arguments passed to the ggplot theme}
}
\value{
ggplot of the discrete or relative length distributions of
clone sequences
}
\description{
This function displays either the nucleotide (\strong{nt}) or amino
acid (\strong{aa}) sequence length. The sequence length visualized
can be selected using the chains parameter, either the combined clone
(both chains) or across all single chains. Visualization can either
be a histogram or if \strong{scale} = TRUE, the output will
be a density plot. Multiple sequencing runs can be group together
using the group.by parameter.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalLength(combined, cloneCall="aa", chain = "both")

}
\concept{Visualizing_Clones}
