% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalLength.R
\name{clonalLength}
\alias{clonalLength}
\title{Demonstrate the distribution of lengths filtered clones}
\usage{
clonalLength(
  df,
  cloneCall = "aa",
  chain = "both",
  group.by = NULL,
  order = TRUE,
  scale = FALSE,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clonotype - CDR3 nucleotide (nt), 
CDR3 amino acid (aa).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The variable to use for grouping.}

\item{order}{Maintain the order of the list when plotting}

\item{scale}{Converts the graphs into density plots in order to show 
relative distributions.}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}
}
\value{
ggplot of the discrete or relative length distributions of 
clonotype sequences
}
\description{
This function displays either the nucleotide (nt) or amino 
acid (aa) sequence length. The sequence length visualized can be 
selected using the chains parameter, either the combined clonotype 
(both chains) or across all single chains. Visualization can either 
be a histogram or if scale = TRUE, the output will be a density plot. 
Multiple sequencing runs can be group together using the 
group parameter.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalLength(combined, cloneCall="aa", chain = "both")

}
\concept{Visualizing_Clones}
