% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{clonalOverlap}
\alias{clonalOverlap}
\title{Examining the clonal overlap between groups or samples}
\usage{
clonalOverlap(
  df,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  method = c("overlap", "morisita"),
  exportTable = F
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), CDR3 nucleotide (nt) or CDR3 amino acid (aa), or
CDR3 gene+nucleotide (gene+nt).}

\item{method}{The method to calculate the overlap, either the overlap coefficient or morisita index}

\item{exportTable}{Exports a table of the data into the global environment in addition to the visualization}
}
\description{
This functions allows for the caclulation and visualizations of the overlap coefficient or morisita index
for clonotypes using the product of combineContig() or expression2list(). The overlap coefficient is calculated
using the intersection of clonotypes divided by the length of the smallest componenet. Morisita index is estimating
the dispersion of a population, more information can be found [here](https://en.wikipedia.org/wiki/Morisita%27s_overlap_index).
If a matrix output for the data is preferred, set exportTable = TRUE.
}
