% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{clonalOverlap}
\alias{clonalOverlap}
\title{Examining the clonal overlap between samples}
\usage{
clonalOverlap(
  df,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  method = c("overlap", "morisita"),
  exportTable = F
)
}
\arguments{
\item{df}{The product of CombineContig() or the seurat object after combineSeurat()}

\item{cloneCall}{How to call the clonotype - CDR3 gene, CDR3 nt or CDR3 aa, or CDR3+nucleotide}

\item{method}{The method to calculate the overlap, either the overlap coefficient or morisita index}

\item{exportTable}{Exports a table of the data into the global environment in addition to the visualization}
}
\description{
Use overlap coefficient or morisita index to exam the overlapping clonotypes across samples. If a matrix output for the data is preferred, set exportTable = T.
}
