% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{clonalOverlap}
\alias{clonalOverlap}
\title{Examining the clonal overlap between groups or samples}
\usage{
clonalOverlap(
  df,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  method = c("overlap", "morisita", "jaccard"),
  chain = "both",
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (gene+nt).}

\item{method}{The method to calculate the overlap, either the overlap 
coefficient, morisita or jaccard indices.}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{exportTable}{Exports a table of the data into the global 
environment in addition to the visualization}
}
\value{
ggplot of the clonotypic overlap between elements of a list
}
\description{
This functions allows for the caclulation and visualizations of the 
overlap coefficient or morisita index for clonotypes using the product 
of combineTCR(), combineBCR() or expression2list(). The overlap 
coefficient is calculated using the intersection of clonotypes 
divided by the length of the smallest componenet. Morisita index is 
estimating the dispersion of a population, more information can be found 
[here](https://en.wikipedia.org/wiki/Morisita%27s_overlap_index).
If a matrix output for the data is preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

clonalOverlap(combined, cloneCall = "gene", method = "overlap")

}
