% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{clonalOverlap}
\alias{clonalOverlap}
\title{Examining the clonal overlap between samples}
\usage{
clonalOverlap(
  df,
  call = c("gene", "nt", "aa", "gene+nt"),
  method = c("overlap", "morisita")
)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{call}{is the how to call the clonotype - CDR3 gene, CDR3 nt or CDR3 aa, or CDR3+nucleotide}

\item{method}{use either the overlap coefficient or morisita index}
}
\description{
Examining the clonal overlap between samples
}
