% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalOverlap.R
\name{clonalOverlap}
\alias{clonalOverlap}
\title{Examining the clonal overlap between groups or samples}
\usage{
clonalOverlap(
  input.data,
  cloneCall = "strict",
  method = NULL,
  chain = "both",
  group.by = NULL,
  order.by = NULL,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{The product of \code{\link{combineTCR}}, 
\code{\link{combineBCR}}, or \code{\link{combineExpression}}}

\item{cloneCall}{How to call the clone - VDJC gene (\strong{gene}), 
CDR3 nucleotide (\strong{nt}), CDR3 amino acid (\strong{aa}),
VDJC gene + CDR3 nucleotide (\strong{strict}) or a custom variable 
in the data}

\item{method}{The method to calculate the "overlap", "morisita", 
"jaccard", "cosine" indices or "raw" for the base numbers}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The variable to use for grouping}

\item{order.by}{A vector of specific plotting order or "alphanumeric"
to plot groups in order}

\item{exportTable}{Returns the data frame used for forming the graph}

\item{palette}{Colors to use in visualization - input any 
\link[grDevices]{hcl.pals}}
}
\value{
ggplot of the overlap of clones by group
}
\description{
This functions allows for the calculation and visualizations of 
various overlap metrics for clones. The methods include overlap 
coefficient (\strong{overlap}), Morisita's overlap index 
(\strong{morisita}), Jaccard index (\strong{jaccard}), cosine 
similarity (\strong{cosine}) or the exact number of clonal 
overlap (\strong{raw}).
}
\details{
The formulas for the indices are as follows:

\strong{Overlap Coefficient:}
\deqn{overlap = \frac{\sum \min(a, b)}{\min(\sum a, \sum b)}}  

\strong{Raw Count Overlap:}
\deqn{raw = \sum \min(a, b)}

\strong{Morisita Index:}
\deqn{morisita = \frac{\sum a b}{(\sum a)(\sum b)}}  

\strong{Jaccard Index:}
\deqn{jaccard = \frac{\sum \min(a, b)}{\sum a + \sum b - \sum \min(a, b)}}  

\strong{Cosine Similarity:}
\deqn{cosine = \frac{\sum a b}{\sqrt{(\sum a^2)(\sum b^2)}}}  

Where:  
\itemize{  
  \item{\eqn{a} and \eqn{b} are the abundances of species \eqn{i} in groups A and B, respectively.}
}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

clonalOverlap(combined, 
              cloneCall = "aa", 
              method = "jaccard")

}
\concept{Visualizing_Clones}
