% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalOverlap.R
\name{clonalOverlap}
\alias{clonalOverlap}
\title{Examining the clonal overlap between groups or samples}
\usage{
clonalOverlap(
  input.data,
  cloneCall = "strict",
  method = NULL,
  chain = "both",
  group.by = NULL,
  exportTable = FALSE,
  palette = "inferno",
  pcoa = FALSE,
  pcoa.group.by = NULL,
  point.size = 3
)
}
\arguments{
\item{input.data}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa),
VDJC gene + CDR3 nucleotide (strict) or a custom variable in the data.}

\item{method}{The method to calculate the "overlap", "morisita", 
"jaccard", "cosine" indices or "raw" for the base numbers.}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The variable to use for grouping.}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}

\item{pcoa}{Plot PCoA plot instead of a heatmap.}

\item{pcoa.group.by}{grouping variable in PCoA plot, default to NULL.}

\item{point.size}{point size in PCoA plot.}
}
\value{
ggplot of the overlap of clones by group
}
\description{
This functions allows for the calculation and visualizations of 
various overlap metrics for clonotypes.
}
\details{
The formulas for the indices are as follows:

\strong{Overlap Coefficient:}
\deqn{overlap = \frac{\sum \min(a_i, b_i)}{\min(\sum a_i, \sum b_i)}}

\strong{Raw Count Overlap:}
\deqn{raw = \sum \min(a_i, b_i)}

\strong{Morisita Index:}
\deqn{morisita = \frac{\sum a_i b_i}{(\sum a_i)(\sum b_i)}}

\strong{Jaccard Index:}
\deqn{jaccard = \frac{\sum \min(a_i, b_i)}{\sum a_i + \sum b_i - \sum \min(a_i, b_i)}}

\strong{Cosine Similarity:}
\deqn{cosine = \frac{\sum a_i b_i}{\sqrt{(\sum a_i^2)(\sum b_i^2)}}}

Where:
\itemize{
  \item{\eqn{a_i} and \eqn{b_i} are the abundances of species \eqn{i} in groups A and B, respectively.}
}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

clonalOverlap(combined, 
              cloneCall = "aa", 
              method = "jaccard")

}
\concept{Visualizing_Clones}
