% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalProportion.R
\name{clonalProportion}
\alias{clonalProportion}
\title{Examining the clonal space occupied by specific clonotypes}
\usage{
clonalProportion(
  df,
  split = c(10, 100, 1000, 10000, 30000, 1e+05),
  cloneCall = "gene+nt",
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of CombineContig() or expression2List()}

\item{split}{The cutpoints for the specific clonotypes.}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), 
CDR3 nucleotide (nt) or CDR3 amino acid (aa), or CDR3 
gene+nucleotide (gene+nt).}

\item{exportTable}{Exports a table of the data into the global 
environment in addition to the visualization}
}
\value{
ggplot of the space occupied by the specific rank of clonotypes
}
\description{
This function calculates the relative clonal space occupied by the 
clonotypes. The grouping of these clonotypes is based on the parameter 
split, at default, split will group the clonotypes into bins of 1:10, 
11:100, 101:1001, etc. To adjust the clonotypes selected, change the 
numbers in the variable split. If a matrix output for the data is
preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")
clonalProportion(combined, cloneCall = "gene")

}
