% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalProportion.R
\name{clonalProportion}
\alias{clonalProportion}
\title{Examining the clonal space occupied by specific clonotypes}
\usage{
clonalProportion(
  df,
  clonal.split = c(10, 100, 1000, 10000, 30000, 1e+05),
  cloneCall = "strict",
  chain = "both",
  group.by = NULL,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{clonal.split}{The cut points for the specific clonotypes.}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL".}

\item{group.by}{The variable to use for grouping.}

\item{exportTable}{Exports a table of the data into the global.
environment in addition to the visualization.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}
}
\value{
ggplot of the space occupied by the specific rank of clonotypes
}
\description{
This function calculates the relative clonal space occupied by the 
clonotypes. The grouping of these clonotypes is based on the parameter 
**clonal.split**, at default, **clonal.split** will group the clonotypes 
into bins of 1:10, 11:100, 101:1001, etc. To adjust the clonotypes 
selected, change the numbers in the variable split. If a matrix output 
for the data is preferred, set exportTable = TRUE.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalProportion(combined, cloneCall = "gene")

}
