% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalProportion.R
\name{clonalProportion}
\alias{clonalProportion}
\title{Examining the clonal space occupied by specific clonotypes}
\usage{
clonalProportion(
  df,
  split = c(10, 100, 1000, 10000, 30000, 1e+05),
  call = c("gene", "nt", "aa", "gene+nt"),
  exportTable = F
)
}
\arguments{
\item{df}{The product of CombineContig() or the seurat object after combineSeurat()}

\item{split}{The cutpoints for the specific clonotypes}

\item{call}{How to call the clonotype - CDR3 gene, CDR3 nt or CDR3 aa, or CDR3+nucleotide}

\item{exportTable}{Exports a table of the data into the global environment in addition to the visualization}
}
\description{
Examining the clonal space occupied by specific clonotypes
}
