% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalScatter.R, R/scatterClonotypes.R
\name{clonalScatter}
\alias{clonalScatter}
\title{Scatter plot comparing the expansion of two samples}
\usage{
clonalScatter(
  df,
  cloneCall = "strict",
  x.axis = NULL,
  y.axis = NULL,
  chain = "both",
  dot.size = "total",
  split.by = NULL,
  graph = "proportion",
  exportTable = FALSE,
  palette = "inferno",
  seed = NULL
)

clonalScatter(
  df,
  cloneCall = "strict",
  x.axis = NULL,
  y.axis = NULL,
  chain = "both",
  dot.size = "total",
  split.by = NULL,
  graph = "proportion",
  exportTable = FALSE,
  palette = "inferno",
  seed = NULL
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{x.axis}{name of the list element to appear on the x.axis}

\item{y.axis}{name of the list element to appear on the y.axis}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{dot.size}{either total or the name of the list element to 
use for size of dots}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{graph}{graph either proportion or raw clonotype count}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any hcl.pals()}

\item{seed}{the integer seed to set for the random variation of point coords.}
}
\value{
ggplot of the relative clonotype numbers

ggplot of the relative clonotype numbers
}
\description{
This function produces a scatter plot directly comparing the specific clonotypes
between two samples. The clonotypes will be categorized by counts into singlets or multiplets, 
either exclusive or shared between the selected samples. Visualization inspired 
by the work of \href{https://pubmed.ncbi.nlm.nih.gov/32103181/}{Wu, T, et al 2020}.

This function produces a scatter plot directly comparing the specific clonotypes
between two samples. The clonotypes will be categorized by counts into singlets or multiplets, 
either exclusive or shared between the selected samples. Visualization inspired 
by the work of \href{https://pubmed.ncbi.nlm.nih.gov/32103181/}{Wu, T, et al 2020}.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalScatter(combined, 
              x.axis = "P17B", 
              y.axis = "P17L",
              graph = "proportion")

#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalScatter(combined, 
              x.axis = "P17B", 
              y.axis = "P17L",
              graph = "proportion")

}
