% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalSizeDistribution.R
\name{clonalSizeDistribution}
\alias{clonalSizeDistribution}
\title{Hierarchical clustering of clonotypes on clonotype size and 
Jensen-Shannon divergence}
\usage{
clonalSizeDistribution(
  df,
  cloneCall = "strict",
  chain = "both",
  method = "ward.D2",
  threshold = 1,
  group.by = NULL,
  split.by = NULL,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), 
expression2List(), or combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene),
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{method}{The clustering parameter for the dendrogram.}

\item{threshold}{Numerical vector containing the thresholds 
the grid search was performed over.}

\item{group.by}{The column header used for grouping.}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any hcl.pals()}
}
\value{
ggplot dendrogram of the clone size distribution
}
\description{
This function produces a hierarchical clustering of clonotypes by sample 
using the Jensen-Shannon distance and discrete gamma-GPD spliced threshold 
model in 
\href{https://bioconductor.org/packages/devel/bioc/html/powerTCR.html}{powerTCR R package}.
Please read and cite PMID: 30485278 if using the function for analyses.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalSizeDistribution(combined, cloneCall = "strict", method="ward.D2")

}
\author{
Hillary Koch
}
