% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clonalSizeDistribution.R
\name{clonalSizeDistribution}
\alias{clonalSizeDistribution}
\title{Hierarchical clustering of clones using Gamma-GPD spliced threshold model}
\usage{
clonalSizeDistribution(
  df,
  cloneCall = "strict",
  chain = "both",
  method = "ward.D2",
  threshold = 1,
  group.by = NULL,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene),
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL".}

\item{method}{The clustering parameter for the dendrogram.}

\item{threshold}{Numerical vector containing the thresholds 
the grid search was performed over.}

\item{group.by}{The variable to use for grouping.}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}
}
\value{
ggplot dendrogram of the clone size distribution
}
\description{
This function produces a hierarchical clustering of clonotypes by sample 
using discrete gamma-GPD spliced threshold 
model. If using this model please read and cite 
\href{https://pubmed.ncbi.nlm.nih.gov/30485278/}{PMID: 30485278}.
}
\details{
The probability density function (pdf) for the Generalized Pareto Distribution (GPD) is given by:
\deqn{f(x|\mu, \sigma, \xi) = \frac{1}{\sigma} \left(1 + \xi \left( \frac{x - \mu}{\sigma} \right) \right)^{-\left( \frac{1}{\xi} + 1 \right)}}{
f(x|mu, sigma, xi) = 1/sigma * (1 + xi * ((x - mu)/sigma))^(-1/xi - 1)
}

Where:
\itemize{
  \item{\eqn{\mu} is a location parameter}
  \item{\eqn{\sigma > 0} is a scale parameter}
  \item{\eqn{\xi} is a shape parameter}
  \item{\eqn{x \ge \mu} if \eqn{\xi \ge 0} and \eqn{\mu \le x \le \mu - \sigma/\xi} if \eqn{\xi < 0}}
}

The probability density function (pdf) for the Gamma Distribution is given by:
\deqn{f(x|\alpha, \beta) = \frac{x^{\alpha-1} e^{-x/\beta}}{\beta^\alpha \Gamma(\alpha)}}{
f(x|alpha, beta) = (x^(alpha-1) * exp(-x/beta)) / (beta^alpha * Gamma(alpha))
}

Where:
\itemize{
  \item{\eqn{\alpha > 0} is the shape parameter}
  \item{\eqn{\beta > 0} is the scale parameter}
  \item{\eqn{x \ge 0}}
  \item{\eqn{\Gamma(\alpha)} is the gamma function of \eqn{\alpha}}
}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
clonalSizeDistribution(combined, cloneCall = "strict", method="ward.D2")

}
\author{
Hillary Koch
}
