% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterTCR.R
\name{clusterTCR}
\alias{clusterTCR}
\title{Clustering T cell receptors}
\usage{
clusterTCR(
  df,
  chain = NULL,
  sequence = NULL,
  threshold = 0.85,
  group.by = NULL
)
}
\arguments{
\item{df}{The product of combineTCR(), expression2List(), or combineExpression().}

\item{chain}{The TCR to cluster - TRA, TRB, TRG, TRD}

\item{sequence}{Clustering based on either "aa" or "nt"}

\item{threshold}{The normalized edit distance to consider. The higher the number the more 
similarity of sequence will be used for clustering.}

\item{group.by}{The column header used for to group contigs.}
}
\value{
List of clonotypes for individual cell barcodes
}
\description{
This function uses edit distances of either the nucleotide or amino acid 
sequences of the CDR3 and V genes to cluster similar TCRs together. 
The distance clustering will then be amended to the end of the list of 
combined contigs. The cluster will appear as CHAIN.num if a unique sequence or 
CHAIN:LD.num if clustered together.T his function will only two chains 
recovered, multiple chains will automatically be reduced. This function 
also underlies the combineBCR() function and therefore not needed for 
B cells. This may take some time to calculate the distances and cluster.
}
\examples{
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))

sub_combined <- clusterTCR(combined[[2]], chain = "TRA", sequence = "aa")

}
