% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{combineBCR}
\alias{combineBCR}
\title{Combining the list of B Cell Receptor contigs}
\usage{
combineBCR(df, samples = NULL, ID = NULL, removeNA = F, removeMulti = F)
}
\arguments{
\item{df}{List of filtered contig annotations from 10x Genomics}

\item{samples}{The labels of samples}

\item{ID}{The additional sample labeling option}

\item{removeNA}{This will remove any chain without values}

\item{removeMulti}{This will remove barcodes with greater than 2 chains}
}
\description{
Combining the output of the 10x Genomics filtered BCR contig annotation files into a list of processed contigs. Selection of the samples and ID characters should match the prefix of the seurat object if attaching is planned. The function incorporates the calculation of normalized Hamming's distance for the estimation of clonotype,
allowing for the identification of cells with >= 0.85 in nucleotide sequence. This function will assign clonotypes to the nucleotide sequences sequentially, unless the Hamming's Distance Criteria is met, and will be modified with an HD. The calculation is performed across all barcodes, regardless of the sample. The strict clonotype definition will add the correponding vgene as well.
}
