% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{combineBCR}
\alias{combineBCR}
\title{Combining the list of B cell receptor contigs into clones}
\usage{
combineBCR(
  input.data,
  samples,
  ID = NULL,
  call.related.clones = TRUE,
  threshold = 0.85,
  removeNA = FALSE,
  removeMulti = FALSE,
  filterMulti = TRUE,
  filterNonproductive = TRUE
)
}
\arguments{
\item{input.data}{List of filtered contig annotations or outputs from
\code{\link[=loadContigs]{loadContigs()}}.}

\item{samples}{The labels of samples (required).}

\item{ID}{The additional sample labeling (optional).}

\item{call.related.clones}{Use the nucleotide sequence and V gene
to call related clones. Default is set to TRUE. FALSE will return
a CTstrict or strict clone as V gene + amino acid sequence.}

\item{threshold}{The normalized edit distance to consider. The higher
the number the more similarity of sequence will be used for clustering.}

\item{removeNA}{This will remove any chain without values.}

\item{removeMulti}{This will remove barcodes with greater than 2 chains.}

\item{filterMulti}{This option will allow for the selection of the
highest-expressing light and heavy chains, if not calling related clones.}

\item{filterNonproductive}{This option will allow for the removal of
nonproductive chains if the variable exists in the contig data. Default
is set to TRUE to remove nonproductive contigs.}
}
\value{
List of clones for individual cell barcodes
}
\description{
This function consolidates a list of BCR sequencing results to the level
of the individual cell barcodes. Using the samples and ID parameters,
the function will add the strings as prefixes to prevent issues with
repeated barcodes. The resulting new barcodes will need to match the
Seurat or SCE object in order to use, \code{\link[=combineExpression]{combineExpression()}}.
Unlike \code{\link[=combineTCR]{combineTCR()}}, combineBCR produces a column
\strong{CTstrict} of an index of nucleotide sequence and the
corresponding V gene. This index automatically calculates the
Levenshtein distance between sequences with the same V gene and will
index sequences using a normalized Levenshtein distance with the same
ID. After which, clone clusters are called using the
\code{\link[igraph:components]{igraph::components()}} function. Clones that are clustered
across multiple sequences will then be labeled with "Cluster" in the
CTstrict header.
}
\examples{
#Data derived from the 10x Genomics intratumoral NSCLC B cells
BCR <- read.csv("https://www.borch.dev/uploads/contigs/b_contigs.csv")
combined <- combineBCR(BCR, 
                       samples = "Patient1", 
                       threshold = 0.85)

}
\concept{Loading_and_Processing_Contigs}
