% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{combineContigs}
\alias{combineContigs}
\title{Combining the list of contigs}
\usage{
combineContigs(df, samples = NULL, ID = NULL, cells = c("T-AB", "T-GD", "B"))
}
\arguments{
\item{df}{List of sontigs}

\item{samples}{are the labels of samples}

\item{ID}{is the additional sample lebeling option}

\item{cells}{is the type of lymphocyte - Tcell-AB, T-Cell-GD, or B cell}
}
\description{
Combining the list of contigs
}
