% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineExpression.R
\name{combineExpression}
\alias{combineExpression}
\title{Adding clonotype information to a single-cell object}
\usage{
combineExpression(
  df,
  sc,
  cloneCall = "strict",
  chain = "both",
  group.by = "none",
  proportion = TRUE,
  filterNA = FALSE,
  cloneSize = c(Rare = 1e-04, Small = 0.001, Medium = 0.01, Large = 0.1, Hyperexpanded =
    1),
  addLabel = FALSE
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}} or a list of 
both c(\code{\link{combineTCR}}, c\code{\link{combineBCR}})}

\item{sc}{The seurat or SingleCellExperiment (SCE) object to attach}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt) CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The column label in the combined clones in which 
clonotype frequency will be calculated. "none" will keep the list as is, 
while NULL will merge all the clones into a single data frame.}

\item{proportion}{Whether to proportion (**TRUE**) or total frequency (**FALSE**) 
of the clonotype based on the group.by variable.}

\item{filterNA}{Method to subset seurat object of barcodes without 
clonotype information}

\item{cloneSize}{The bins for the grouping based on proportion or frequency. 
If proportion is **FALSE** and the *cloneSizes* are not set high enough based 
on frequency, the upper limit of *cloneSizes* will be automatically amended.}

\item{addLabel}{This will add a label to the frequency header, allowing
the user to try multiple group.by variables or recalculate frequencies after 
subsetting the data.}
}
\value{
seurat or SingleCellExperiment object with attached clonotype 
information
}
\description{
This function adds the immune receptor information to the Seurat or 
SCE object to the meta data. By default this function also calculates 
the frequencies and proportion of the clonotypes by sequencing 
run (**group.by** = "none"). To change how the frequencies/proportions
are calculated, select a column header for the **group.by** variable. 
Importantly, before using \code{\link{combineExpression}} ensure the 
barcodes of the single-cell object object match the barcodes in the output 
of the \code{\link{combineTCR}} or \code{\link{combineBCR}}.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

#Getting a sample of a Seurat object
scRep_example <- get(data("scRep_example"))
sce <- suppressMessages(Seurat::UpdateSeuratObject(scRep_example))
sce <- Seurat::as.SingleCellExperiment(sce)

#Using combineExpresion()
sce <- combineExpression(combined, sce)

}
\concept{SC_Functions}
