% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{combineExpression}
\alias{combineExpression}
\title{Adding clonotype information to a Seurat or SCE object}
\usage{
combineExpression(
  df,
  sc,
  cloneCall = "strict",
  chain = "both",
  group.by = "none",
  proportion = TRUE,
  filterNA = FALSE,
  cloneTypes = c(Rare = 1e-04, Small = 0.001, Medium = 0.01, Large = 0.1, Hyperexpanded =
    1),
  addLabel = FALSE
)
}
\arguments{
\item{df}{The product of CombineTCR() or CombineBCR() or a list of 
both c(CombineTCR(), combineBCR())}

\item{sc}{The seurat or SingleCellExperiment (SCE) object to attach}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt) CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The column label in the combined contig object in which 
clonotype frequency will be calculated. "none" will keep the list as is, 
while NULL will merge all the contigs into a single data frame.}

\item{proportion}{Whether to use the total frequency (FALSE) or the 
proportion (TRUE) of the clonotype based on the group.by variable.}

\item{filterNA}{Method to subset seurat object of barcodes without 
clonotype information}

\item{cloneTypes}{The bins for the grouping based on frequency}

\item{addLabel}{This will add a label to the frequency header, allowing
the user to try multiple group.by variables or recalculate frequencies after 
subseting the data.}
}
\value{
seurat or SingleCellExperiment object with attached clonotype 
information
}
\description{
This function adds the immune receptor information to the Seurat or 
SCE object to the meta data. By default this function also calculates 
the frequencies of the clonotypes by sequencing run (group.by = "none"). 
To change how the frequencies are calculated, select a column header for 
the group.by variable. Importantly, before using combineExpression() 
ensure the barcodes of the seurat or SCE object match the barcodes in the 
output of the combinedContig() call. Check changeNames() to change the 
prefix of the Seurat object. If combining more than one immune receptor 
type, barcodes with both receptors will be removed during the combination 
process.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
sce <- suppressMessages(Seurat::UpdateSeuratObject(screp_example))
sce <- Seurat::as.SingleCellExperiment(sce)

#Using combineExpresion()
sce <- combineExpression(combined, sce)

}
