% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{combineTCR}
\alias{combineTCR}
\title{Combining the list of T Cell Receptor contigs}
\usage{
combineTCR(
  df,
  samples = NULL,
  ID = NULL,
  cells = c("T-AB", "T-GD"),
  removeNA = F,
  removeMulti = F,
  filterNA = F
)
}
\arguments{
\item{df}{List of filtered contig annotations from 10x Genomics}

\item{samples}{The labels of samples}

\item{ID}{The additional sample labeling option}

\item{cells}{The type of lymphocyte - T cell-AB or T cell-GD}

\item{removeNA}{This will remove any chain without values}

\item{removeMulti}{This will remove barcodes with greater than 2 chains}

\item{filterNA}{This option will allow for the selection of the 2 corresponding chains with the highest expression for a single barcode}
}
\description{
Combining the output of the 10x Genomics filtered TCR contig annotation files into a list of processed contigs. Selection of the samples and ID characters should match the prefix of the seurat object if attaching is planned. If more than 2 productive chains exist for a barcode, the sequences will be combined and incorporated into a clonotype.
}
