% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{combineTCR}
\alias{combineTCR}
\title{Combining the list of T Cell Receptor contigs}
\usage{
combineTCR(
  df,
  samples = NULL,
  ID = NULL,
  removeNA = FALSE,
  removeMulti = FALSE,
  filterMulti = FALSE
)
}
\arguments{
\item{df}{List of filtered contig annotations from 10x Genomics.}

\item{samples}{The labels of samples (required).}

\item{ID}{The additional sample labeling (optional).}

\item{removeNA}{This will remove any chain without values.}

\item{removeMulti}{This will remove barcodes with greater than 2 chains.}

\item{filterMulti}{This option will allow for the selection of the 2 
corresponding chains with the highest expression for a single barcode.}
}
\value{
List of clonotypes for individual cell barcodes
}
\description{
This function consolidates a list of TCR sequencing results to the level of 
the individual cell barcodes. Using the samples and ID parameters, the 
function will add the strings as prefixes to prevent issues with repeated 
barcodes. The resulting new barcodes will need to match the Seurat or SCE 
object in order to use, \code{\link{combineExpression}}. Several 
levels of filtering exist - remove or filterMulti are parameters that 
control how  the function  deals with barcodes with multiple chains 
recovered.
}
\examples{
combineTCR(contig_list, 
           samples = rep(c("PX", "PY", "PZ"), each=2), 
           ID = rep(c("P", "T"), 3))

}
