% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{compareClonotypes}
\alias{compareClonotypes}
\title{Demonstrate the difference in clonal proportion between clonotypes}
\usage{
compareClonotypes(
  df,
  cloneCall = "strict",
  chain = "both",
  samples = NULL,
  clonotypes = NULL,
  numbers = NULL,
  split.by = NULL,
  graph = "alluvial",
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{samples}{The specific samples to isolate for visualization.}

\item{clonotypes}{The specific sequences of interest.}

\item{numbers}{The top number clonotype sequences per group}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{graph}{The type of graph produced, either "alluvial" or "area".}

\item{exportTable}{Returns the data frame used for forming the graph.}
}
\value{
ggplot of the proportion of total sequencing read of 
selecting clonotypes
}
\description{
This function produces an alluvial or area graph of the proportion of 
the indicated clonotypes for all or selected samples. Clonotypes can be 
selected using the clonotypes parameter with the specific sequence of 
interest or using the number parameter with the top n clonotypes by 
proportion to be visualized.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))
compareClonotypes(combined, numbers = 10, 
samples = c("PX_P", "PX_T"), cloneCall="aa")

}
