% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contig_list}
\alias{contig_list}
\title{A List of Eight Single-cell TCR Sequencing Runs.}
\description{
A list of 8 \code{filtered_contig_annotations.csv} files
outputted from 10X Cell Ranger. More information on the
data can be found in the following
\href{https://pubmed.ncbi.nlm.nih.gov/33622974/}{manuscript}.
}
\concept{Data}
