% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contig_list.R
\docType{data}
\name{contig_list}
\alias{contig_list}
\title{A data set of T cell contigs as a list outputed from the 
filter_contig_annotation files.}
\description{
A data set of T cell contigs as a list outputed from the 
filter_contig_annotation files.
}
