% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{createHTOContigList}
\alias{createHTOContigList}
\title{Generate a contig list from a multiplexed experiment}
\usage{
createHTOContigList(contigs, sc, group.by = NULL)
}
\arguments{
\item{sc}{The Seurat or SCE object.}

\item{contig}{The filtered contig annotation file from multiplexed experiment}
}
\value{
Returns a list of contigs corresponding to the multiplexed Seurat 
or Single-Cell Experiment object
}
\description{
Multiplexing single-cell sequencing runs is an efficient method for quantifying 
multiple samples or conditions simultaneously. Unfortunately, the hashing information
is not stored in the TCR sequence data. In order preprocess and form a contig list
for downstream analysis in scRepertoire, createHTOContigList() take the filtered contig 
annotation output and the single-cell RNA object to create the list.
}
