% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportClones.R
\name{exportClones}
\alias{exportClones}
\title{Exporting clonotypes}
\usage{
exportClones(
  df,
  group.by = NULL,
  split.by = NULL,
  write.file = TRUE,
  dir = NULL,
  file.name = "clones.csv"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{group.by}{The column header used for grouping.}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{write.file}{TRUE, save the file, FALSE, return a data.frame}

\item{dir}{directory location to save the csv}

\item{file.name}{the csv file name}
}
\value{
ggplot of percentage of V and J gene pairings as a heatmap
}
\description{
This function saves a csv file of clonotypes (genes, amino acid, and nucleotide sequences)
by barcodes.
}
\examples{
\dontrun{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
exportClones(combined)
}
                                   
}
\author{
Jonathan Noonan, Nick Borcherding
}
