% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportClones.R
\name{exportClones}
\alias{exportClones}
\title{Exporting clonotypes}
\usage{
exportClones(
  input.data,
  format = "paired",
  group.by = NULL,
  write.file = TRUE,
  dir = NULL,
  file.name = "clones.csv"
)
}
\arguments{
\item{input.data}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{format}{The format to export the clones - "paired", "airr", or "TCRMatch".}

\item{group.by}{The variable to use for grouping.}

\item{write.file}{\strong{TRUE}, save the file or \strong{FALSE}, return a data.frame}

\item{dir}{directory location to save the csv}

\item{file.name}{the csv file name}
}
\value{
CSV file of the paired sequences.
}
\description{
This function saves a csv file of clonotypes (genes, amino acid, and nucleotide sequences)
by barcodes. \strong{format} determines the structure of the csv file - \emph{paired} will
export sequences by barcodes and include multiple chains, \emph{airr} will export a data 
frame that is consistent with the AIRR format, and \emph{TCRMatch} will 
export a data frame that has the TRB chain with count information.
}
\examples{
\dontrun{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
exportClones(combined, 
             format = "paired")
}
                                   
}
\author{
Jonathan Noonan, Nick Borcherding
}
\concept{Loading_and_Processing_Contigs}
