% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression2List.R
\name{expression2List}
\alias{expression2List}
\title{Allows users to take the meta data in Seurat/SCE and place it into a list 
that will work with all the functions}
\usage{
expression2List(sc, split.by)
}
\arguments{
\item{sc}{object after combineExpression().}

\item{split.by}{The column header to group the new list. NULL will return clusters.}
}
\value{
list derived from the meta data of single-cell object with 
elements divided by the group parameter
}
\description{
Allows users to perform more fundamental measures of clonotype analysis 
using the meta data from the Seurat or SCE object. For Seurat objects the 
active identity is automatically added as "cluster". Remaining grouping 
parameters or SCE or Seurat objects must appear in the meta data.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

#Getting a sample of a Seurat object
scRep_example <- get(data("scRep_example"))
scRep_example <- combineExpression(combined, scRep_example)

#Using expression2List
newList <- expression2List(scRep_example, split.by = "seurat_clusters")

}
