% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{expression2List}
\alias{expression2List}
\title{\emph{\strong{DEPRECATED}} Take the meta data in seurat/SCE and place it into a list}
\usage{
expression2List(sc, ..., force = FALSE)
}
\arguments{
\item{sc}{output of \code{\link[=combineExpression]{combineExpression()}}.}

\item{...}{previously the \code{group} or \code{split.by} argument, indicating the
column header to group the new list by. This should strictly be one argument
and is an ellipsis for backwards compatibility. Everything after the first
argument is ignored.}

\item{force}{logical. If \code{TRUE}, the function will run with a deprecation
warning. Else throws a deprecation error.}
}
\value{
list derived from the meta data of single-cell object with
elements divided by the group parameter
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Allows users to perform more fundamental measures of clonotype analysis
using the meta data from the seurat or SCE object. For Seurat objects the
active identity is automatically added as "cluster". Remaining grouping
parameters or SCE or Seurat objects must appear in the meta data.

This function is deprecated as of version 2 due to the confusion it caused
to many users. Users are encouraged to remain with the abstraction barrier
of combined single cell objects and the outputs of \code{\link[=combineTCR]{combineTCR()}} and
\code{\link[=combineBCR]{combineBCR()}} for all functions.

We discourage the use of this function, but if you have to use it, set the
\code{force} argument to \code{TRUE}.
}
\keyword{internal}
