% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircilize.R
\name{getCirclize}
\alias{getCirclize}
\title{Generate data frame to be used with circlize R package to visualize
clonotypes as a chord diagram.}
\usage{
getCirclize(sc, cloneCall = "strict", group.by = NULL, proportion = FALSE)
}
\arguments{
\item{sc}{object after combineExpression().}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{group.by}{The group header for which you would like to analyze 
the data.}

\item{proportion}{Binary will calculate relationship unique 
clonotypes (proportion = FALSE) or a ratio of the group.by 
variable (proportion = TRUE)}
}
\value{
data frame of shared clonotypes between groups
}
\description{
This function will take the meta data from the product of 
combineExpression()and generate a relational data frame to 
be used for a chord diagram. Each cord will represent the number of 
clonotype unqiue and shared across the multiple group.by variable.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

#Getting a sample of a Seurat object
scRep_example <- get(data("scRep_example"))
scRep_example <- combineExpression(combined, scRep_example)

#Getting data frame output for Circilize
circles <- getCirclize(scRep_example, group.by = "seurat_clusters")


}
\author{
Dillon Corvino, Nick Borcherding
}
