% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{getCirclize}
\alias{getCirclize}
\title{Generate data frame to be used with circlize R package to visualize
clonotypes as a chord diagram.}
\usage{
getCirclize(sc, cloneCall = "gene+nt", groupBy = NULL, proportion = FALSE)
}
\arguments{
\item{sc}{object after combineExpression().}

\item{cloneCall}{How to call the clonotype - CDR3 nucleotide (nt), 
CDR3 amino acid (aa).}

\item{groupBy}{The group header for which you would like to analyze 
the data.}

\item{proportion}{Binary will calculate relationship as unique clonotypes 
(proportion = TRUE) or proportion of unique clonotypes (proportion = FALSE)}
}
\value{
data frame of shared clonotypes between groups
}
\description{
This function will take the meta data from the product of combineExpression()
and generate a relational data frame to be used for a chord diagram. The output
is a measure of relative clonotype overlap between groups and does not reflect
exact clonotype matches between groups.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
screp_example <- combineExpression(combined, screp_example)

#Getting data frame output for Circilize
circles <- getCirclize(screp_example, groupBy = "seurat_clusters")


}
\author{
Dillon Corvino, Nick Borcherding
}
