% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCircilize.R
\name{getCirclize}
\alias{getCirclize}
\title{Generate data frame to be used with circlize R package to visualize
clones as a chord diagram.}
\usage{
getCirclize(sc.data, cloneCall = "strict", group.by = NULL, proportion = FALSE)
}
\arguments{
\item{sc.data}{The single-cell object after \code{\link{combineExpression}}.}

\item{cloneCall}{How to call the clone - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa),
VDJC gene + CDR3 nucleotide (strict) or a custom variable in the data.}

\item{group.by}{The group header for which you would like to analyze 
the data.}

\item{proportion}{Calculate the relationship unique 
clones (proportion = FALSE) or normalized by 
proportion (proportion = TRUE)}
}
\value{
A data frame of shared clones between groups formated for \link[circlize]{chordDiagram}
}
\description{
This function will take the meta data from the product of 
\code{\link{combineExpression}} and generate a relational data frame to 
be used for a chord diagram. Each cord will represent the number of 
clone unique and shared across the multiple \strong{group.by} variable. 
If using the downstream circlize R package, please read and cite the
following \href{https://pubmed.ncbi.nlm.nih.gov/24930139/}{manuscript}.
If looking for more advance ways for circular visualizations, there
is a great \href{https://jokergoo.github.io/circlize_book/book/}{cookbook}
for the circlize package.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

#Getting a sample of a Seurat object
scRep_example <- get(data("scRep_example"))
scRep_example <- combineExpression(combined, 
                                   scRep_example)

#Getting data frame output for Circlize
circles <- getCirclize(scRep_example, 
                       group.by = "seurat_clusters")


}
\author{
Dillon Corvino, Nick Borcherding
}
\concept{SC_Functions}
