% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combineContigs.R
\name{hammingCompare}
\alias{hammingCompare}
\title{Calculates the normalized Hamming Distance between the contig nucleotide sequence.}
\usage{
hammingCompare(Con.df, gene, chain, length)
}
\arguments{
\item{gene}{The IGH or IG light chains (IGLC)}

\item{chain}{The column header with the nucletoide sequence}

\item{length}{The column header with the specific length}
}
\value{
#Normalized Hammings Distance comparisons to be used in combineBCR()
}
\description{
This feature caluclates the normalized Hammings Distance, or the Hammings Distance 
divided by length of sequence in order to index similar and divergent sequences 
involved in combineBCR(). It is not designed as an indepenent function. The 
threshold for similar sequences is set to > 0.85 normalized Hammings Distance, 
which if met will index the similar sequence into a single sequence and add 
":HD" to the index.
}
\keyword{internal}
