% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlightClones.R
\name{highlightClones}
\alias{highlightClones}
\title{Highlighting specific clonotypes in Seurat}
\usage{
highlightClones(
  sc,
  cloneCall = c("gene", "nt", "aa", "strict"),
  sequence = NULL
)
}
\arguments{
\item{sc}{The single-cell object to attach after \code{\link{combineExpression}}}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{sequence}{The specific sequence or sequence to highlight}
}
\value{
Single-cell object object with new meta data column for indicated clones
}
\description{
Use a specific clonotype sequence to highlight on top of the dimensional 
reduction in single-cell object.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

#Getting a sample of a Seurat object
scRep_example  <- get(data("scRep_example"))

#Using combineExpresion()
scRep_example  <- combineExpression(combined, scRep_example)

#Using highlightClonotype()
scRep_example   <- highlightClones(scRep_example , cloneCall= "aa", 
sequence = c("CVVSDNTGGFKTIF_CASSVRRERANTGELFF"))

}
