% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{highlightClonotypes}
\alias{highlightClonotypes}
\title{Highlighting specific clonotypes in Seurat}
\usage{
highlightClonotypes(
  seurat,
  call = c("gene", "nt", "aa", "gene+nt"),
  sequence = NULL
)
}
\arguments{
\item{seurat}{The seurat object to attach}

\item{call}{How to call the clonotype - based on genes, CDR3 nt, or CDR3 aa sequence or the combination of genes and nucleotide sequence}

\item{sequence}{The specifc sequence or sequence to highlight}
}
\description{
Highlighting specific clonotypes in Seurat
}
