% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{highlightClonotypes}
\alias{highlightClonotypes}
\title{Highlighting specific clonotypes in Seurat}
\usage{
highlightClonotypes(
  sc,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  sequence = NULL
)
}
\arguments{
\item{sc}{The seurat object to attach}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
CDR3 gene+nucleotide (gene+nt).}

\item{sequence}{The specifc sequence or sequence to highlight}
}
\value{
DimPlot with highlighted clonotypes
}
\description{
Use a specific clonotype sequence to highlight on top of the dimensional 
reduction in seurat object.
}
\examples{
#' #Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

#Getting a sample of a Seurat object
URL <- "https://ncborcherding.github.io/vignettes/seurat_example.rda"
dataDest = file.path("seurat_example.rda")
download.file(URL, destfile = dataDest, method = "auto")
seurat <- load("seurat_example.rda")

#Using combineExpresion()
seurat <- combineExpression(combined, seurat)

#Using highlightClonotype()
seurat <- highlightClonotypes(seurat, cloneCall= "aa", 
sequence = c("CAVNGGSQGNLIF_CSAEREDTDTQYF")

}
