% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{highlightClonotypes}
\alias{highlightClonotypes}
\title{Highlighting clonotypes in Seurat}
\usage{
highlightClonotypes(
  seurat,
  call = c("gene", "nt", "aa", "gene+nt"),
  sequence = NULL
)
}
\arguments{
\item{seurat}{is the seurat object to attach}

\item{call}{is the how to call the clonotype - based on genes, CDR3 nt, or CDR3 aa sequence}

\item{sequence}{the specifc sequence or sequence to highlight}
}
\description{
Highlighting clonotypes in Seurat
}
