% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{highlightClonotypes}
\alias{highlightClonotypes}
\title{Highlighting specific clonotypes in Seurat}
\usage{
highlightClonotypes(
  sc,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  sequence = NULL
)
}
\arguments{
\item{sc}{The seurat object to attach}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), CDR3 nucleotide (nt) or CDR3 amino acid (aa), or
CDR3 gene+nucleotide (gene+nt).}

\item{sequence}{The specifc sequence or sequence to highlight}
}
\value{
DimPlot with highlighted clonotypes
}
\description{
Use a specific clonotype sequence to highlight on top of the dimensional reduction in seurat object.
}
\examples{
highlightClonotype(seurat, cloneCall = "gene", sequence = "CAVNGGSQGNLIF_CSAEREDTDTQYF")

}
