% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{highlightClonotypes}
\alias{highlightClonotypes}
\title{Highlighting specific clonotypes in Seurat}
\usage{
highlightClonotypes(
  sc,
  cloneCall = c("gene", "nt", "aa", "strict"),
  sequence = NULL
)
}
\arguments{
\item{sc}{The Seurat object to attach}

\item{cloneCall}{How to call the clonotype - VDJC gene (gene), 
CDR3 nucleotide (nt), CDR3 amino acid (aa), or 
VDJC gene + CDR3 nucleotide (strict).}

\item{sequence}{The specific sequence or sequence to highlight}
}
\value{
Seurat object with new meta data column for indicated clones
}
\description{
Use a specific clonotype sequence to highlight on top of the dimensional 
reduction in seurat object.
}
\examples{
#' #Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))

#Using combineExpresion()
screp_example <- combineExpression(combined, screp_example )

#Using highlightClonotype()
screp_example  <- highlightClonotypes(screp_example, cloneCall= "aa", 
sequence = c("CAVNGGSQGNLIF_CSAEREDTDTQYF"))

}
