% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{lengthContig}
\alias{lengthContig}
\title{Demonstrate the distribution of lengths filtered contigs output from 10x Genomics}
\usage{
lengthContig(df, call = c("nt", "aa"), column = NULL, scale = F)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{call}{is the how to call the clonotype - CDR3 nt or CDR3 aa sequence.}

\item{column}{is the column header for which you would like to analyze the data}

\item{scale}{converts the graphs into denisty plots in order to show relative distributions.}
}
\description{
Demonstrate the distribution of lengths filtered contigs output from 10x Genomics
}
