% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{lengthContig}
\alias{lengthContig}
\title{Demonstrate the distribution of lengths filtered contigs.}
\usage{
lengthContig(
  df,
  cloneCall = "aa",
  chain = "both",
  group.by = NULL,
  split.by = NULL,
  order = TRUE,
  scale = FALSE,
  exportTable = FALSE
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), expression2List(), or combineExpression().}

\item{cloneCall}{How to call the clonotype - CDR3 nucleotide (nt), 
CDR3 amino acid (aa).}

\item{chain}{indicate if both or a specific chain should be used - 
e.g. "both", "TRA", "TRG", "IGH", "IGL"}

\item{group.by}{The group header for which you would like to analyze 
the data.}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{order}{Maintain the order of the list when plotting}

\item{scale}{Converts the graphs into density plots in order to show 
relative distributions.}

\item{exportTable}{Returns the data frame used for forming the graph.}
}
\value{
ggplot of the discrete or relative length distributions of 
clonotype sequences
}
\description{
This function takes the output of combineTCR(), combineBCR(), or
expression2List() and displays either the nucleotide (nt) or amino 
acid (aa) sequence length. The sequence length visualized can be 
selected using the chains parameter, either the combined clonotype 
(both chains) or across all single chains. Visualization can either 
be a histogram or if scale = TRUE, the output will be a density plot. 
Multiple sequencing runs can be group together using the 
group parameter. If a matrix output for the data is preferred, set 
exportTable = TRUE.
}
\examples{
#Making combined contig data
x <- contig_list
combined <- combineTCR(x, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))
lengthContig(combined, cloneCall="aa", chain = "both")

}
