% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{occupiedscRepertoire}
\alias{occupiedscRepertoire}
\title{Visualize the number of single cells with clonotype frequencies by cluster}
\usage{
occupiedscRepertoire(
  sc,
  x.axis = "ident",
  label = TRUE,
  facet.by = NULL,
  proportion = FALSE,
  na.include = FALSE,
  exportTable = FALSE
)
}
\arguments{
\item{sc}{The Seurat or SCE object to visualize after combineExpression().}

\item{x.axis}{The variable in the meta data to graph along the x.axis}

\item{label}{Include the number of clonotype in each category by x.axis variable}

\item{facet.by}{The column header used for faceting the graph}

\item{proportion}{Convert the stacked bars into relative proportion}

\item{na.include}{Visualize NA values or not.}

\item{exportTable}{Exports a table of the data into the global 
environment in addition to the visualization}
}
\value{
Stacked bar plot of counts of cells by clonotype frequency group
}
\description{
View the count of clonotypes frequency group in seurat or SCE object 
meta data after combineExpression(). The visualization will take the 
new meta data variable "cloneType" and plot the number of cells with
each designation using a secondary variable, like cluster. Credit to 
the idea goes to Drs. Carmona and Andreatta and their work with
\href{https://github.com/carmonalab/ProjecTILs}{ProjectTIL}.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
sce <- suppressMessages(Seurat::UpdateSeuratObject(screp_example))

#Using combineExpresion()
sce <- combineExpression(combined, sce)

#Using occupiedscRepertoire()
occupiedscRepertoire(sce, x.axis = "ident")
table <- occupiedscRepertoire(sce, x.axis = "ident", exportTable = TRUE)

}
