% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentAA.R
\name{percentAA}
\alias{percentAA}
\title{Examining the relative amino acid composition by position}
\usage{
percentAA(
  df,
  chain = "TRB",
  group.by = NULL,
  split.by = NULL,
  aa.length = 20,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of combineTCR(), combineBCR(), 
expression2List(), or combineExpression().}

\item{chain}{"TRA", "TRB", "TRG", "TRG", "IGH", "IGL"}

\item{group.by}{The column header used for grouping.}

\item{split.by}{If using a single-cell object, the column header 
to group the new list. NULL will return clusters.}

\item{aa.length}{The maximum length of the cdr3 amino acid sequence}

\item{exportTable}{Returns the data frame used for forming the graph}

\item{palette}{Colors to use in visualization - input any hcl.pals()}
}
\value{
ggplot of stacked bar graphs of amino acid proportions
}
\description{
This function the proportion of amino acids along the residues 
of the cdr3 amino acid sequence.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
percentAA(combined, chain = "TRB", aa.length = 20)
}
