% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentGenes.R
\name{percentGenes}
\alias{percentGenes}
\title{Examining the V or J gene usage across groupings}
\usage{
percentGenes(
  df,
  chain = "TRB",
  gene = "Vgene",
  group.by = NULL,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{chain}{"TRA", "TRB", "TRG", "TRG", "IGH", "IGL"}

\item{gene}{"V", "D" or "J"}

\item{group.by}{The variable to use for grouping.}

\item{exportTable}{Returns the data frame used for forming the graph}

\item{palette}{Colors to use in visualization - input any hcl.pals()}
}
\value{
ggplot of percentage of indicated genes as a heatmap
}
\description{
This function the proportion V or J genes used by 
grouping variables
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
percentGenes(combined, chain = "TRB", gene = "Vgene")

}
