% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentGeneUsage.R
\name{percentGeneUsage}
\alias{percentGeneUsage}
\alias{vizGenes}
\alias{percentGenes}
\alias{percentVJ}
\title{Summarizes and Visualizes Gene Usage}
\usage{
percentGeneUsage(
  input.data,
  chain = "TRB",
  genes = "TRBV",
  group.by = NULL,
  order.by = NULL,
  summary.fun = c("percent", "proportion", "count"),
  plot.type = "heatmap",
  exportTable = FALSE,
  palette = "inferno"
)

vizGenes(
  input.data,
  x.axis = "TRBV",
  y.axis = NULL,
  group.by = NULL,
  plot = "heatmap",
  order.by = NULL,
  summary.fun = c("percent", "proportion", "count"),
  exportTable = FALSE,
  palette = "inferno"
)

percentGenes(
  input.data,
  chain = "TRB",
  gene = "Vgene",
  group.by = NULL,
  order.by = NULL,
  exportTable = FALSE,
  summary.fun = c("percent", "proportion", "count"),
  palette = "inferno"
)

percentVJ(
  input.data,
  chain = "TRB",
  group.by = NULL,
  order.by = NULL,
  summary.fun = c("percent", "proportion", "count"),
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{The product of \code{\link[=combineTCR]{combineTCR()}}, \code{\link[=combineBCR]{combineBCR()}},
or \code{\link[=combineExpression]{combineExpression()}}.}

\item{genes}{A character vector specifying the gene loci to analyze.
Can be a single gene e.g., "TRBV" or "IGHJ" or a pair for genes analysis
(e.g., c("TRBV", "TRAV"), or "TRBV", "TRBJ").}

\item{group.by}{The variable to use for grouping (e.g., "sample", "orig.ident").}

\item{order.by}{A vector of specific plotting order or "alphanumeric"
to plot groups in order.}

\item{summary.fun}{Character string choosing the summary statistic -
\code{"percent"} (default), \code{"proportion"}, or \code{"count"}.}

\item{plot.type}{The type of plot to return: \code{"heatmap"} (default for paired loci,
also available for single loci), or \code{"barplot"} (for single loci).}

\item{exportTable}{Logical. If \code{TRUE}, returns the data frame or matrix used
for forming the graph instead of the plot.}

\item{palette}{Colors to use in visualization - input any
\link[grDevices:palettes]{hcl.pals}.}
}
\value{
A ggplot object displaying a heatmap or bar plot of gene usage.
If \code{exportTable = TRUE}, a matrix or data frame of the raw data is returned.
}
\description{
This function quantifies and visualizes the usage of V, D, or J genes,
or gene pairings across T or B cell clones.
}
\concept{Summarize_Repertoire}
\concept{Visualizing_Clones}
