% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentKmer.R
\name{percentKmer}
\alias{percentKmer}
\title{Examining the relative composition of kmer motifs}
\usage{
percentKmer(
  df,
  chain = "TRB",
  cloneCall = "aa",
  group.by = NULL,
  motif.length = 3,
  top.motifs = 30,
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{chain}{"TRA", "TRB", "TRG", "TRG", "IGH", "IGL".}

\item{cloneCall}{How to call the clonotype - CDR3 nucleotide (nt) or 
CDR3 amino acid (aa).}

\item{group.by}{The variable to use for grouping.}

\item{motif.length}{The length of the kmer to analyze.}

\item{top.motifs}{Return the n most variable motifs as a function of 
median absolute deviation.}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}
}
\value{
ggplot of percentage of kmers as a heatmap
}
\description{
This function the of kmer for nucleotide or amino acids 
in the CDR3 sequence.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
percentKmer(combined, chain = "TRB", motif.length = 3)
}
