% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percentVJ.R
\name{percentVJ}
\alias{percentVJ}
\title{Examining the V or J gene usage across groupings}
\usage{
percentVJ(
  input.data,
  chain = "TRB",
  group.by = NULL,
  exportTable = FALSE,
  palette = "inferno",
  pcoa = FALSE,
  pcoa.group.by = NULL,
  dist.method = "manhattan",
  point.size = 3
)
}
\arguments{
\item{input.data}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{chain}{"TRA", "TRB", "TRG", "TRG", "IGH", "IGL"}

\item{group.by}{The variable to use for grouping.}

\item{exportTable}{Returns the data frame used for forming the graph}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}

\item{pcoa}{Plot PCoA plot instead of a heatmap.}

\item{pcoa.group.by}{grouping variable in PCoA plot, default to NULL.}

\item{dist.method}{distance calculation method to use in dist().}

\item{point.size}{point size in PCoA plot.}
}
\value{
ggplot of percentage of V and J gene pairings as a heatmap
}
\description{
This function the proportion V or J genes used by 
grouping variables
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
percentVJ(combined, chain = "TRB")

}
\concept{Summarize_Repertoire}
