% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalEntropy.R
\name{positionalEntropy}
\alias{positionalEntropy}
\title{Examining the diversity of amino acids by position}
\usage{
positionalEntropy(
  input.data,
  chain = "TRB",
  group.by = NULL,
  order.by = NULL,
  aa.length = 20,
  method = "norm.entropy",
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{The product of \code{\link[=combineTCR]{combineTCR()}},
\code{\link[=combineBCR]{combineBCR()}}, or \code{\link[=combineExpression]{combineExpression()}}}

\item{chain}{indicate a specific chain should be used -
e.g."TRA", "TRG", "IGH", "IGL", etc}

\item{group.by}{The variable to use for grouping}

\item{order.by}{A vector of specific plotting order for \code{group.by} or
"alphanumeric" to plot groups in order}

\item{aa.length}{The maximum length of the CDR3 amino acid sequence.}

\item{method}{The method to calculate the entropy/diversity -
\code{"shannon"}, \code{"inv.simpson"}, \code{"gini.simpson"}, \code{"norm.entropy"},
\code{"pielou"}, \code{"hill0"}, \code{"hill1"}, \code{"hill2"}}

\item{exportTable}{Returns the data frame used for forming the graph}

\item{palette}{Colors to use in visualization - input any \link[grDevices:palettes]{hcl.pals}}
}
\value{
ggplot of line graph of diversity by amino acid residue position
}
\description{
This function the diversity amino acids along the residues
of the CDR3 amino acid sequence. Please see
\code{\link[=clonalDiversity]{clonalDiversity()}} for more information on
the underlying methods for diversity/entropy calculations.
Positions without variance will have a value reported as 0
for the purposes of comparison.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
positionalEntropy(combined, 
                  chain = "TRB", 
                  aa.length = 20)
}
\concept{Summarize_Repertoire}
