% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positionalProperty.R
\name{positionalProperty}
\alias{positionalProperty}
\title{Examining the mean property of amino acids by position}
\usage{
positionalProperty(
  input.data,
  chain = "TRB",
  group.by = NULL,
  aa.length = 20,
  method = "Atchley",
  exportTable = FALSE,
  palette = "inferno"
)
}
\arguments{
\item{input.data}{The product of \code{\link{combineTCR}}, 
\code{\link{combineBCR}}, or \code{\link{combineExpression}}.}

\item{chain}{"TRA", "TRB", "TRG", "TRG", "IGH", "IGL".}

\item{group.by}{The variable to use for grouping.}

\item{aa.length}{The maximum length of the CDR3 amino acid sequence.}

\item{method}{The method to calculate the property - "Atchley", "Kidera",
"stScales", "tScales", or "VHSE".}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}
}
\value{
ggplot of line graph of diversity by position
}
\description{
This function calculates the mean selected property for 
amino acids along the residues of the CDR3 amino acid sequence. 
The ribbon surrounding the individual line represents the 95% 
confidence interval.
}
\details{
More information for the individual methods can be found at the following citations:

\strong{Atchley:} \href{https://pubmed.ncbi.nlm.nih.gov/15851683/}{citation}

\strong{Kidera:} \href{https://link.springer.com/article/10.1007/BF01025492}{citation}

\strong{stScales:} \href{https://pubmed.ncbi.nlm.nih.gov/19373543/}{citation}

\strong{tScales:} \href{https://www.sciencedirect.com/science/article/pii/S0022286006006314?casa_token=uDj97DwXDDEAAAAA:VZfahldPRwU1WObySJlohudtMSDwF7nJSUzcEGwPhvkY13ALLKhs08Cf0_FyyfYZjxJlj-fVf0SM}{citation}

\strong{VHSE:} \href{https://pubmed.ncbi.nlm.nih.gov/15895431/}{citation}
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
positionalProperty(combined, 
                   chain = "TRB",
                   method = "Atchley" 
                   aa.length = 20)
}
\concept{Summarize_Repertoire}
