% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{quantContig}
\alias{quantContig}
\title{Quantify the unique clonotypes in the filtered contigs.}
\usage{
quantContig(
  df,
  cloneCall = c("gene", "nt", "aa", "gene+nt"),
  scale = F,
  group = NULL,
  exportTable = F
)
}
\arguments{
\item{df}{The product of combineContig().}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), CDR3 nucleotide (nt) or CDR3 amino acid (aa), or
CDR3 gene+nucleotide (gene+nt).}

\item{scale}{Converts the graphs into percentage of unique clonotypes.}

\item{group}{The column header used for grouping.}

\item{exportTable}{Returns the data frame used for forming the graph,}
}
\description{
This function takes the output from combineContig() or expression2List() and quantifies unique clonotypes.
The unique clonotypes can be either reported as a raw output or scaled to the total number of clonotypes recovered
using the scale parameter. Multiple sequencing runs can be group together using the group parameter. If a matrix
output for the data is preferred, set exportTable = TRUE.
}
