% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regressClonotype.R
\name{regressClonotype}
\alias{regressClonotype}
\title{Normalize feature space of single-cell expression by clonotype}
\usage{
regressClonotype(
  sc,
  cloneCall = "gene+nt",
  groupVariable = NULL,
  reduction = NULL,
  dims = NULL,
  ...
)
}
\arguments{
\item{sc}{The seurat or SCE object to visualize after combineExpression(). 
For SCE objects, the cluster variable must be in the meta data under 
"cluster".}

\item{cloneCall}{How to call the clonotype - CDR3 gene (gene), 
CDR3 nucleotide (nt) or CDR3 amino acid (aa), or 
CDR3 gene+nucleotide (gene+nt).}

\item{groupVariable}{Additional features to normalize the single cell object with}

\item{reduction}{Re-perform the dimensional reduction - "UMAP" or "TSNE"}

\item{dims}{The number of dimensions to use in the reduction (i.e., 1:20)}
}
\value{
Seurat or SingleCellExperiment object
}
\description{
This function reduces the influence of clonotype on the feature space of the 
single-cell object. It first removes the genes that comprise the adaptive immune 
receptors from the list of variable genes. Next, it uses the 
\href{https://github.com/immunogenomics/harmony}{harmony R package} to re-integrate 
data to minimize effects of clonotype and other variables indicated by groupVariable.
}
\examples{
#Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3), cells ="T-AB")

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
sce <- suppressMessages(Seurat::UpdateSeuratObject(screp_example))
sce <- Seurat::as.SingleCellExperiment(sce)

#Using combineExpresion()
sce <- combineExpression(combined, sce)

sce <- regressClonotype(sce, groupVariable = "patient", reduction = "UMAP", dims = 1:20)


}
