% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRep_example}
\alias{scRep_example}
\title{A seurat object of 100 single T cells derived
from 3 clear cell renal carcinoma patients.}
\description{
The object is compatible with `contig_list` and the TCR
sequencing data can be added with `combineExpression`.
}
