% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRep_example}
\alias{scRep_example}
\title{A Seurat object of 500 single T cells,}
\description{
The object is compatible with `contig_list` and the TCR
sequencing data can be added with `combineExpression`.  The data is 
from 4 patients with acute respiratory distress, with samples taken
from both the lung and peripheral blood. More information on the 
data can be found in the following
 \href{https://pubmed.ncbi.nlm.nih.gov/33622974/}{manuscript}.
}
\concept{Data}
