% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scRep_example}
\alias{scRep_example}
\title{A seurat object of 500 single T cells derived
from 4 patients with acute respiratory distress (ARDS).}
\description{
The object is compatible with `contig_list` and the TCR
sequencing data can be added with `combineExpression`.
}
