% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRepertoire-package.R
\docType{package}
\name{scRepertoire-package}
\alias{scRepertoire}
\alias{scRepertoire-package}
\title{scRepertoire: A toolkit for single-cell immune receptor profiling}
\description{
scRepertoire was built to process data derived from the 10x Genomics Chromium Immune Profiling for both T-cell receptor (TCR) and immunoglobulin (Ig) enrichment workflows and subsequently interacts with the popular Seurat and SingleCellExperiment R packages. It also allows for general analysis of single-cell clonotype information without the use of expression information. The package functions as a wrapper for Startrac and powerTCR R packages.
}
\author{
\strong{Maintainer}: Nick Borcherding \email{ncborch@gmail.com}

}
\keyword{internal}
