% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRepertoire-package.R
\docType{package}
\name{scRepertoire-package}
\alias{scRepertoire}
\alias{scRepertoire-package}
\title{scRepertoire: A toolkit for single-cell immune receptor profiling}
\description{
scRepertoire is a toolkit for processing and analyzing single-cell T-cell receptor (TCR) and immunoglobulin (Ig). The scRepertoire framework supports use of 10x, AIRR, BD, MiXCR, Omniscope, TRUST4, and WAT3R single-cell formats. The functionality includes basic clonal analyses, repertoire summaries, distance-based clustering and interaction with the popular Seurat and SingleCellExperiment/Bioconductor R workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.borch.dev/uploads/scRepertoire/}
  \item Report bugs at \url{https://github.com/BorchLab/scRepertoire/issues}
}

}
\author{
\strong{Maintainer}: Nick Borcherding \email{ncborch@gmail.com}

Authors:
\itemize{
  \item Qile Yang \email{qile.yang@berkeley.edu}
  \item Ksenia Safina \email{safina@broadinstitute.org}
}

}
\keyword{internal}
