% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seuratFunctions.R
\name{StartracDiversity}
\alias{StartracDiversity}
\title{Diversity indices for single-cell RNA-seq}
\usage{
StartracDiversity(sc, type = "Type", sample = NULL, by = c("overall"))
}
\arguments{
\item{sc}{The Seurat or SCE object. For SCE objects, the cluster variable must be in the meta data under "cluster".}

\item{type}{The column header in the meta data that gives the where the cells were derived from, not the patient
sample IDs.}

\item{sample}{The column header corresponding to individual samples or patients.}

\item{by}{Method to subset the indices by either overall (across all samples) or by specific group.}
}
\value{
ggplot of clonal metrics calculatrd by the Startrac R package.
}
\description{

}
