% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startracDiversity.R
\name{StartracDiversity}
\alias{StartracDiversity}
\title{Calculate Startrac-based Diversity Indices}
\usage{
StartracDiversity(
  sc.data,
  cloneCall = "strict",
  chain = "both",
  type = NULL,
  group.by = NULL,
  exportTable = FALSE,
  palette = "inferno",
  ...
)
}
\arguments{
\item{sc.data}{The single-cell object after \code{\link[=combineExpression]{combineExpression()}}.
For SCE objects, the cluster variable must be in the meta data under
"cluster".}

\item{cloneCall}{Defines the clonal sequence grouping. Accepted values
are: \code{gene} (VDJC genes), \code{nt} (CDR3 nucleotide sequence), \code{aa} (CDR3 amino
acid sequence), or \code{strict} (VDJC). A custom column header can also be used.}

\item{chain}{The TCR/BCR chain to use. Use \code{both} to include both chains
(e.g., TRA/TRB). Accepted values: \code{TRA}, \code{TRB}, \code{TRG}, \code{TRD}, \code{IGH}, \code{IGL}
(for both light chains), \code{both}.}

\item{type}{The variable in the meta data that provides tissue type.}

\item{group.by}{A column header in the metadata or lists to group the analysis
by (e.g., "sample", "treatment"). If \code{NULL}, data will be analyzed as
by list element or active identity in the case of single-cell objects.}

\item{exportTable}{If \code{TRUE}, returns a data frame or matrix of the results
instead of a plot.}

\item{palette}{Colors to use in visualization - input any \link[grDevices:palettes]{hcl.pals}.}

\item{...}{Additional arguments passed to the ggplot theme}
}
\value{
A ggplot object visualizing STARTRAC diversity metrics or data.frame if
\code{exportTable = TRUE}.
}
\description{
This function utilizes the Startrac approach derived from
\href{https://pubmed.ncbi.nlm.nih.gov/30479382/}{PMID: 30479382}.
Required to run the function, the "type" variable needs to include the
difference in where the cells were derived. The output of this function
will produce 3 indices: \code{expa} (clonal expansion), \code{migra}
(cross-tissue migration), and \code{trans} (state transition). In order
to understand the underlying analyses of the outputs please
read and cite the linked manuscript.
}
\examples{
# Getting the combined contigs
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

# Getting a sample of a Seurat object
scRep_example  <- get(data("scRep_example"))
scRep_example  <- combineExpression(combined, scRep_example)
scRep_example$Patient <- substring(scRep_example$orig.ident,1,3)
scRep_example$Type <- substring(scRep_example$orig.ident,4,4) 

# Using StartracDiversity()
StartracDiversity(scRep_example, 
                  type = "Type", 
                  group.by = "Patient")

}
\author{
Liangtao Zheng
}
\concept{SC_Functions}
