% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startrac.R
\name{StartracDiversity}
\alias{StartracDiversity}
\title{Diversity indices for single-cell RNA-seq}
\usage{
StartracDiversity(
  sc,
  type = "Type",
  sample = NULL,
  by = "overall",
  exportTable = FALSE
)
}
\arguments{
\item{sc}{The seurat or SCE object to visualize after combineExpression(). 
For SCE objects, the cluster variable must be in the meta data under 
"cluster".}

\item{type}{The column header in the meta data that gives the where the 
cells were derived from, not the patient sample IDs}

\item{sample}{The column header corresponding to individual samples 
or patients.}

\item{by}{Method to subset the indices by either overall 
(across all samples) or by specific group}

\item{exportTable}{Returns the data frame used for forming the graph}
}
\value{
ggplot object of Startrac diversity metrics
}
\description{
This function utilizes the Startrac R package derived from 
\href{https://pubmed.ncbi.nlm.nih.gov/30479382/}{PMID: 30479382}
Required to run the function, the "type" variable needs to include the 
difference in where the cells were derived. The output of this function 
will produce 3 indices: expa (clonal expansion), migra 
(cross-tissue migration), and trans (state transition). In order 
to understand the underlying analyses of the outputs please 
read and cite the linked manuscript.
}
\examples{
\dontrun{ 
#Getting the combined contigs
combined <- combineTCR(contig_list, rep(c("PX", "PY", "PZ"), each=2), 
rep(c("P", "T"), 3))

#Getting a sample of a Seurat object
screp_example <- get(data("screp_example"))
screp_example <- combineExpression(combined, screp_example)

#Using occupiedscRepertoire()
StartracDiversity(screp_example, type = "Type", sample = "Patient", by = "overall")
}

}
