% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{stripBarcode}
\alias{stripBarcode}
\title{Removing any additional prefixes to the barcodes of filtered contigs.}
\usage{
stripBarcode(contigs, column = 1, connector = "_", num_connects = 3)
}
\arguments{
\item{contigs}{The raw loaded filtered_contig_annotation.csv}

\item{column}{The column in which the barcodes are listed}

\item{connector}{The type of character in which is attaching the defualt 
barcode with any other characters}

\item{num_connects}{The number of strings combined with the connectors}
}
\value{
list with the suffixes of the barcodes removed.
}
\description{
Removing any additional prefixes to the barcodes of filtered contigs.
}
\examples{
stripBarcode(contig_list[[1]], column = 1, connector = "_", num_connects = 1)
}
