% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetContig.R
\name{subsetContig}
\alias{subsetContig}
\title{Subset the product of combineTCR() combineBCR()}
\usage{
subsetContig(df, name, variables = NULL)
}
\arguments{
\item{df}{The product of \code{\link{combineTCR}} or \code{\link{combineBCR}}.}

\item{name}{The column header you'd like to use to subset.}

\item{variables}{The values to subset by, must be in the names(df).}
}
\value{
list of contigs that have been filtered for the name parameter
}
\description{
This function allows for the subsetting of the product of 
\code{\link{combineTCR}} or \code{\link{combineBCR}} 
by the name of the individual list element.
}
\examples{
x <- contig_list
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))
subset <- subsetContig(combined, name = "sample", variables = c("P17B"))

}
