% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{subsetContig}
\alias{subsetContig}
\title{Adding variables aftering the combination of contigs.}
\usage{
subsetContig(df, name, variables = NULL)
}
\arguments{
\item{df}{The product of CombineContig()}

\item{name}{The column header you'd like to use to subset}

\item{variables}{The values to subset by}
}
\description{
Adding variables aftering the combination of contigs.
}
