% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{subsetContig}
\alias{subsetContig}
\title{Subset the combineContig() product.}
\usage{
subsetContig(df, name, variables = NULL)
}
\arguments{
\item{df}{The product of CombineContig().}

\item{name}{The column header you'd like to use to subset.}

\item{variables}{The values to subset by, must be in the names(df).}
}
\description{
This function allows for the subsetting of the product of combineContig() by the name of the individual list element.
In gneral the names of are samples + _ + ID, allowing for users to subset the product of combineContig() across a
string or indivudal name.
}
