% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vizGenes.R
\name{vizGenes}
\alias{vizGenes}
\title{Visualizing the distribution of gene usage}
\usage{
vizGenes(
  input.data,
  x.axis = "TRBV",
  y.axis = NULL,
  group.by = NULL,
  plot = "heatmap",
  order = "gene",
  scale = TRUE,
  exportTable = FALSE,
  palette = "inferno",
  pcoa.group.by = NULL,
  dist.method = "manhattan",
  point.size = 3
)
}
\arguments{
\item{input.data}{The product of \code{\link{combineTCR}}, \code{\link{combineBCR}}, or
\code{\link{combineExpression}}.}

\item{x.axis}{Gene segments to separate the x-axis, such as "TRAV", "TRBD", "IGKJ".}

\item{y.axis}{Variable to separate the y-axis, can be both categorical or other gene 
gene segments, such as "TRAV", "TRBD", "IGKJ".}

\item{group.by}{Variable in which to group the diversity calculation.}

\item{plot}{The type of plot to return - heatmap or barplot}

\item{order}{Categorical variable to organize the x-axis, either "gene" or "variance"}

\item{scale}{Converts the individual count of genes to proportion using the total 
respective repertoire size}

\item{exportTable}{Returns the data frame used for forming the graph.}

\item{palette}{Colors to use in visualization - input any \link[grDevices]{hcl.pals}.}

\item{pcoa.group.by}{grouping variable in PCoA plot, default to NULL.}

\item{dist.method}{distance calculation method to use in dist().}

\item{point.size}{point size in PCoA plot.}
}
\value{
ggplot bar diagram or heatmap of gene usage
}
\description{
This function will allow for the visualizing the distribution 
of the any VDJ and C gene of the TCR or BCR using heatmap or
bar chart. This function requires assumes two chains were used in 
defining clonotype, if not, it will default to the only chain 
present regardless of the chain parameter.
}
\examples{
#Making combined contig data
combined <- combineTCR(contig_list, 
                        samples = c("P17B", "P17L", "P18B", "P18L", 
                                    "P19B","P19L", "P20B", "P20L"))

vizGenes(combined, 
         x.axis = "TRBV",
         y.axis = NULL,
         plot = "heatmap")

}
\concept{Visualizing_Clones}
